+++ .gitignore
@@ -1,2 +1,2 @@
 # Created by venv; see https://docs.python.org/3/library/venv.html
-*
+#*
