+++ zx-next-unite.py
@@ -297,104 +297,13 @@ from PySide6.QtWidgets import (
 )
 
 import rc_backgrounds
+# --- Extracted modules (refactored out of this file) ------------------
+from zxnu_config import *
+from zxnu_workers import *
+from zxnu_media import *
+from zxnu_gallery import *
+# ----------------------------------------------------------------------
 
-ZX_NEXT_UNITE_VERSION = "6.7"
-# Set to False to hide all Download / Send to SD Card / Send via NextSync
-# buttons and context-menu actions for the respective pane.
-ZX_NEXT_UNITE_ZXDB_ENABLE_DOWNLOAD_BUTTONS  = False
-ZX_NEXT_UNITE_ZXART_ENABLE_DOWNLOAD_BUTTONS = False
-ZX_NEXT_UNITE_SHOW_ZXDB_PANE = True
-ZX_NEXT_UNITE_SHOW_ZXART_PANE = True
-
-ZX_NEXT_UNITE_ICON_IMAGE_FILE = "zx-next-unite.png"
-ZX_NEXT_UNITE_VERBOSE_LOG_MODE = False
-ZX_NEXT_UNITE_UI_SIZE_MULTIPLIER = 1
-ZX_NEXT_UNITE_UI_WIDTH = 900 * ZX_NEXT_UNITE_UI_SIZE_MULTIPLIER
-ZX_NEXT_UNITE_UI_HEIGTH = 650 * ZX_NEXT_UNITE_UI_SIZE_MULTIPLIER
-ZX_NEXT_UNITE_CONFIG_FILE_NAME = os.path.join(os.path.dirname(os.path.abspath(sys.argv[0])), "hdfg.cfg")
-ZX_NEXT_UNITE_TAB_TITLE_GOOEY = "TOOL: SD Card Utility"
-ZX_NEXT_UNITE_TAB_TITLE_NEXTSYNC = "TOOL: Network Transfer Manager (NextSync)"
-ZX_NEXT_UNITE_TAB_TITLE_NEXTSYNC_SYNCON = "NextSync - Sync ON"
-ZX_NEXT_UNITE_TAB_TITLE_GETIT = "🌍 GetIt"
-ZX_NEXT_UNITE_TAB_TITLE_ZXDB  = "🌍 ZXDB/ZXinfo.dk"
-ZX_NEXT_UNITE_TAB_TITLE_ZXART = "🌍 ZXArt.ee"
-ZX_NEXT_UNITE_TAB_TITLE_FAVORITES = "🌍 ♥ Favorites"
-ZX_NEXT_UNITE_TAB_TITLE_ALLINONE  = "🌍 Unite!"
-
-GETIT_BASE_URL = "https://zxnext.uk"
-GETIT_USER_AGENT = f"ZX-Next-Unite/{ZX_NEXT_UNITE_VERSION}"
-GETIT_PAGE_SIZE = 18
-
-# Minimum number of characters required before a keyword search is accepted.
-SEARCH_MIN_CHARS = 3
-
-ZXDB_BASE_URL = "https://api.zxinfo.dk/v3"
-ZXDB_USER_AGENT = f"ZX-Next-Unite/{ZX_NEXT_UNITE_VERSION}"
-ZXDB_PAGE_SIZE = 20
-
-ZXART_BASE_URL = "https://zxart.ee/api"
-ZXART_USER_AGENT = f"ZX-Next-Unite/{ZX_NEXT_UNITE_VERSION}"
-ZXART_PAGE_SIZE = 20
-
-# AllInOne pane aggregates results from GetIt + ZXDB + zxArt. Paging is
-# applied client-side on the merged result list.
-ALLINONE_PAGE_SIZE = GETIT_PAGE_SIZE + ZXDB_PAGE_SIZE + ZXART_PAGE_SIZE
-
-
-HDF_MONKEY_WINDOWS_URL = "https://uto.speccy.org/downloads/hdfmonkey_windows.zip"
-
-SETTING_HDDFILE = "hddffile"
-SETTING_EXPLORERPATH = "explorerpath"
-SETTING_SCREENSIZE = "screensize"
-SETTING_SOUND = "sound"
-SETTING_VSYNC = "vsync"
-SETTING_HERTZ = "hertz"
-SETTING_JOYSTICK = "joy"
-SETTING_CSPECT = "cspect"
-SETTING_CUSTOM = "custom"
-SETTING_ESC = "esc"
-SETTING_NEXTSYNC_EXPLORERPATH = "nextsync_explorerpath"
-SETTING_NEXTSYNC_SYNCONCE = "nextsync_synconce"
-SETTING_NEXTSYNC_ALWAYSSYNC = "nextsync_alwayssync"
-SETTING_NEXTSYNC_SLOWTRANSFER = "nextsync_slowtransfer"
-SETTING_DEFAULT_TAB_WHEN_OPENING = "default_tab"
-SETTING_WARN_IMAGE_NEARLY_FULL = "warn_image_nearly_full"
-SETTING_NO_PROMPT_ON_DELETION  = "no_prompt_on_deletion"
-SETTING_AVAIL_CHECK            = "avail_check"
-SETTING_MULTI_SEARCH           = "multi_search"
-SETTING_SEARCH_AUTOCOMPLETE    = "search_autocomplete"
-SETTING_CRASH_LOG_ENABLED      = "crash_log_enabled"
-SETTING_ZXDB_LAST_MODE         = "zxdb_last_mode"
-SETTING_ZXDB_LAST_QUERY        = "zxdb_last_query"
-SETTING_ZXART_LAST_MODE        = "zxart_last_mode"
-SETTING_ZXART_LAST_QUERY       = "zxart_last_query"
-SETTING_COLOR_UP_DIRECTORY = "color_up_directory"
-SETTING_COLOR_DIR_NAME    = "color_dir_name"
-SETTING_COLOR_DIR_TYPE    = "color_dir_type"
-SETTING_COLOR_FILE_NAME   = "color_file_name"
-SETTING_COLOR_FILE_EXT    = "color_file_ext"
-SETTING_COLOR_FILE_SIZE   = "color_file_size"
-SETTING_IMAGE_HISTORY     = "image_history"
-SETTING_BG_OPACITY        = "bg_opacity"
-SETTING_BG_IMAGE          = "bg_image"          # "" = Random, else filename (basename only)
-SETTING_CONTENT_DISCLAIMER_AGREED = "content_disclaimer_agreed"
-SETTING_GALLERY_ANIM_MODE      = "gallery_anim_mode"        # "hover" (default) or "timer"
-SETTING_GALLERY_ROWS_PER_PAGE  = "gallery_rows_per_page"    # int 1..10, default 2
-SETTING_GALLERY_COLS           = "gallery_cols"             # int: 2 | 4 (default) | 8
-SETTING_GALLERY_IMG_SIZE       = "gallery_img_size"         # "small" | "medium" (default) | "large"
-SETTING_GETIT_VIEW_MODE        = "getit_view_mode"          # "table" (default) or "gallery"
-SETTING_ZXDB_VIEW_MODE         = "zxdb_view_mode"
-SETTING_ZXART_VIEW_MODE        = "zxart_view_mode"
-SETTING_ZXART_LANGUAGE         = "zxart_language"          # "eng" | "pol" | "spa"
-SETTING_FAVORITES              = "favorites"               # JSON list of favorite entries
-SETTING_FAVORITES_VIEW_MODE    = "favorites_view_mode"     # "gallery" (default) or "table"
-SETTING_ALLINONE_VIEW_MODE     = "allinone_view_mode"      # "gallery" (default) or "table"
-DEFAULT_ZXART_LANGUAGE         = "eng"
-ZXART_LANGUAGE_CHOICES         = (
-    ("English", "eng"),
-    ("Polish",  "pol"),
-    ("Spanish", "spa"),
-)
 
 # ---------------------------------------------------------------------------
 # zxArt legal-status code -> human-readable label.
@@ -404,27 +313,6 @@ ZXART_LANGUAGE_CHOICES         = (
 # table is the source of truth; unknown codes encountered at runtime are
 # memoised in ``ZXART_LEGAL_STATUS_CACHE`` (initialised from the static map)
 # so the same value is translated only once per session.
-ZXART_LEGAL_STATUS_LABELS = {
-    "":            "",
-    "unknown":     "Unknown",
-    "original":    "Original",
-    "rerelease":   "Re-release",
-    "adaptation":  "Adaptation",
-    "localization": "Localization",
-    "mod":         "Modification",
-    "crack":       "Cracked",
-    "mia":         "Missing in action",
-    "corrupted":   "Corrupted",
-    "compilation": "Compilation",
-    "incomplete":  "Incomplete",
-    "demoversion": "Demo version",
-    "forbidden":   "Forbidden by author",
-    "unreleased":  "Unreleased",
-    "recovered":   "Recovered",
-    "illegal":     "Illegal",
-    "allowed":     "Distribution allowed",
-}
-ZXART_LEGAL_STATUS_CACHE = dict(ZXART_LEGAL_STATUS_LABELS)
 
 
 # ----------------------------------------------------------------------
@@ -510,243 +398,14 @@ from PySide6.QtGui import QCursor
 
 
 
-def resource_path(relative_path):
-    if hasattr(sys, "_MEIPASS"):
-        return os.path.join(sys._MEIPASS, relative_path)
-    return os.path.join(os.path.abspath("."), relative_path)
 
 
-def zxart_legal_status_label(code) -> str:
-    """Return a human-readable label for a zxArt ``legalStatus`` code.
-
-    Unknown codes are passed through (prettified) and memoised so they are
-    only formatted once per session.
-    """
-    if code is None:
-        return ""
-    key = str(code).strip()
-    if not key:
-        return ""
-    lk = key.lower()
-    cached = ZXART_LEGAL_STATUS_CACHE.get(lk)
-    if cached is not None:
-        return cached
-    # Unknown code: prettify (e.g. "some_value" -> "Some Value") and cache.
-    label = key.replace("_", " ").replace("-", " ").strip()
-    label = label[:1].upper() + label[1:] if label else key
-    ZXART_LEGAL_STATUS_CACHE[lk] = label
-    return label
 
 # UI translation table for the zxArt pane.  Keys are the English source
 # strings; values map language codes -> localised label.  Strings not
 # present in the table fall back to the source key.
-ZXART_UI_TRANSLATIONS = {
-    # Toolbar / controls
-    "Search":           {"pol": "Szukaj",         "spa": "Buscar"},
-    "Deep":             {"pol": "Głęboko",        "spa": "Profundo"},
-    "Productions":      {"pol": "Produkcje",      "spa": "Producciones"},
-    "By letter":        {"pol": "Wg litery",      "spa": "Por letra"},
-    "Pictures":         {"pol": "Obrazy",         "spa": "Imágenes"},
-    "Page:":            {"pol": "Strona:",        "spa": "Página:"},
-    "< Prev":           {"pol": "< Poprzednia",   "spa": "< Anterior"},
-    "Next >":           {"pol": "Następna >",     "spa": "Siguiente >"},
-    "View:":            {"pol": "Widok:",         "spa": "Vista:"},
-    "Table":            {"pol": "Tabela",         "spa": "Tabla"},
-    "Gallery":          {"pol": "Galeria",        "spa": "Galería"},
-    "Language:":        {"pol": "Język:",         "spa": "Idioma:"},
-    # Detail / metadata field names
-    "Title:":           {"pol": "Tytuł:",         "spa": "Título:"},
-    "Year:":            {"pol": "Rok:",           "spa": "Año:"},
-    "Authors:":         {"pol": "Autorzy:",       "spa": "Autores:"},
-    "Groups:":          {"pol": "Grupy:",         "spa": "Grupos:"},
-    "Produced by:":     {"pol": "Wyprodukowane przez:", "spa": "Producido por:"},
-    "Compo:":           {"pol": "Konkurs:",       "spa": "Concurso:"},
-    "Place:":           {"pol": "Miejsce:",       "spa": "Puesto:"},
-    "Languages:":       {"pol": "Języki:",        "spa": "Idiomas:"},
-    "Legal:":           {"pol": "Status prawny:", "spa": "Estado legal:"},
-    "Description:":     {"pol": "Opis:",          "spa": "Descripción:"},
-    "Type:":            {"pol": "Typ:",           "spa": "Tipo:"},
-    "Rating:":          {"pol": "Ocena:",         "spa": "Valoración:"},
-    "Views:":           {"pol": "Wyświetlenia:",  "spa": "Vistas:"},
-    "Tags:":            {"pol": "Tagi:",          "spa": "Etiquetas:"},
-    "Language:":        {"pol": "Język:",         "spa": "Idioma:"},
-    # Table headers
-    "ID":               {"pol": "ID",             "spa": "ID"},
-    "Title":            {"pol": "Tytuł",          "spa": "Título"},
-    "Year":             {"pol": "Rok",            "spa": "Año"},
-    "Author / Group":   {"pol": "Autor / Grupa",  "spa": "Autor / Grupo"},
-    "Author(s)":        {"pol": "Autor(zy)",      "spa": "Autor(es)"},
-    "Genre / Compo":    {"pol": "Gatunek / Konkurs", "spa": "Género / Concurso"},
-    "Tags":             {"pol": "Tagi",           "spa": "Etiquetas"},
-    # Action buttons / context menu
-    "Download":         {"pol": "Pobierz",        "spa": "Descargar"},
-    "Download File":    {"pol": "Pobierz plik",   "spa": "Descargar archivo"},
-    "Send to SD card":  {"pol": "Wyślij na kartę SD", "spa": "Enviar a tarjeta SD"},
-    "Send via NextSync":{"pol": "Wyślij przez NextSync", "spa": "Enviar por NextSync"},
-    # Status/messages
-    "No preview":       {"pol": "Brak podglądu",  "spa": "Sin vista previa"},
-}
 
 
-def _zxart_tr(text: str) -> str:
-    """Translate a zxArt UI string into the currently-selected language."""
-    if not isinstance(text, str) or not text:
-        return text
-    lang = _zxart_lang()
-    if lang == "eng":
-        return text
-    entry = ZXART_UI_TRANSLATIONS.get(text)
-    if not entry:
-        return text
-    return entry.get(lang, text)
-
-
-DEFAULT_GALLERY_ANIM_MODE      = "hover"
-DEFAULT_GALLERY_ROWS_PER_PAGE  = 2
-DEFAULT_GALLERY_COLS           = 4
-DEFAULT_GALLERY_IMG_SIZE       = "medium"
-GALLERY_COLS                   = 4
-GALLERY_MIN_ROWS               = 1
-GALLERY_MAX_ROWS               = 10
-MAX_ALT_TEXT_LINES             = 5
-MAX_IMAGE_HISTORY         = 10
-
-DEFAULT_COLOR_UP_DIRECTORY = "#ff0000"
-DEFAULT_COLOR_DIR_NAME    = "#0000ff"
-DEFAULT_COLOR_DIR_TYPE    = "#0000ff"
-DEFAULT_COLOR_FILE_NAME   = "#00ff00"
-DEFAULT_COLOR_FILE_EXT    = "#00ff00"
-DEFAULT_COLOR_FILE_SIZE   = "#00ff00"
-
-PORT = 2048    # Port to listen on (non-privileged ports are > 1023)
-VERSION3 = "NextSync3"
-VERSION = "NextSync4"
-IGNOREFILE = "syncignore.txt"
-SYNCPOINT = "syncpoint.dat"
-MAX_PAYLOAD = 1024
-NEXTSYNC_UI_HEIGTH_MULTIPLIER = 1
-NEXTSYNC_UI_HEIGTH = 300 * ZX_NEXT_UNITE_UI_SIZE_MULTIPLIER
-IGNOREFILE_DEFAULT_CONTENT = (("syncignore.txt"), ("syncpoint.dat"), ("zx-next-unite.png"),("*.bak"), ("*.py"), ("*.pyproj"), ("*.pyproj"), ("hdfmonkey.exe"), ("hdfg.cfg"))
-
-INIT_LOG = (("NextSync - by Jari Komppa"), ("HDF Monkey - by Matt Westcott"), ("CSpect - by Mike Dailly http://cspect.org"), ("Inspired by HDFM-GOOEY - by em00k"), ("zx-next-unite - by Julien Clauzel 2024"))
-INIT_HELP = ((f"Welcome to zx-next-unite {ZX_NEXT_UNITE_VERSION} help"),
-             (""),
-             ("Introduction:"),
-             ("--------"),
-             ("HdfmGooey was initialy created by emOOk and NextSync by Jari Komppa."),
-             ("A while back I rambled with the idea of an all in one bootstrapper transfer tool to"),
-             ("avoid manipulating SD cards for the Spectrum Next and that was the initial idea of it."),
-             ("Last but not the least some source code was lost from HDFM Gooey and the tool was stuck back in that time,"),
-             ("with the agreement of emOOk I started a rewrite in Python and later with Jari"),
-             ("The point of using Python that would also provide MacOS and Linux portability."),
-             ("Here we are now you have it!"),
-             (""),
-             (""),
-             ("Third party license"),
-             ("-------------------"),
-             ("zx-next-unite is a Qt Application using pyside6 in Python on top of Qt6, which retains the GPLv2 Licensing."),
-             ("Please refer to the LICENSE file on github: https://github.com/jclauzel/zx-next-unite/blob/master/LICENSE.txt."),
-             (""),
-             ("Pyside6 is not bundled and needs to be installed separately (see installation instructions)."),
-             (""),
-             ("Setup & How to:"),
-             ("---------------"),
-             ("Checkout main setup & demo video avaible at: https://youtu.be/FJG-Z0DCIjQ"),
-             ("NextSync Head Over Heels demo: https://www.youtube.com/watch?v=D3_WqTPvjOE"),
-             ("NextSync Night Knight demo: https://www.youtube.com/watch?v=eN1eMIqMCm4"),
-             (""),
-             ("hdfmonkey:"),
-             ("----------"),
-             ("Is a required external component developped by Matt Westcott  that allows to browse the image."),
-             ("You will need to install it to get this application up and fully running."),
-             (""),
-             ("If you are running the app on Windows and hdfmonkey in not present in the same directory,"),
-             ("you will see an error message in the main log Windows as it is missing."),
-             (""),
-             ("If that is the case you will see a 'Download and Install button' bottom right,"),
-             ("once clicked it will try to fetch https://uto.speccy.org/downloads/hdfmonkey_windows.zip "),
-             ("and unzip hdfmonkey executable in the same directory."),
-             ("If the above automated install is successful, you should then be able to select an image and navigate it."),
-             (""),
-             ("On Mac/Linux you will need to install hdfmonkey manually based on the instructions for your platform that can be found at: https://github.com/gasman/hdfmonkey"),
-             (""),
-             ("NextSync:"),
-             ("---------"),
-             ("zx-next-unite implements the <Server> side code and protocol of NextSync by Jari Komppa."),
-             ("It does not require any dot .sync modification and it uses the same very close python logic as nextsync.py."),
-             (""),
-             ("Initial realease on specnext: https://www.specnext.com/forum/viewtopic.php?f=17&t=1715&fbclid=IwAR1njrmr-wEU0DndAxBjO64K_NwY0E2zbqJVaVfiytHE2-A0eL8HWYeDKf8"),
-             ("As a result you will need to run the dot same .sync command on your Next as with the console version and the same network protocol."),
-             (""),
-             ("The latest release v1.2 of the .sync command can be found here https://github.com/Threetwosevensixseven/specnext/releases/tag/nextsync_v1.2 ."),
-             (""),
-             ("You may follow the same instructions as the provided in the readme.txt of that release."),
-             ("On your Spectrum Next, clone or image copy the SYNC command that is located in the above release zip file into your next dot folder."),
-             ("Navigate to NextSync tab, select the root folder to sync on the left."),
-             ("Once you have selected the folder hit the 'prepare sync' button, check the Next Sync log Window on the right."),
-             ("First time you will run .sync on your will be prompter to select the <server> IP address, this machine running NextSync."),
-             ("From the log window pick the IP address from this machine you want to use and type it on your next."),
-             ("Then start the sync server on this maching using the Yes, start sync button and then run the .sync command on your Next."),
-             ("At this point your Spectrum Next will connect to your machine using a network socket and the files will be sent to your next."),
-             ("As it is your Next that will connect to this machine check your firewall alows inbound calls to this machine on port: 2048 by default." ),
-             (""),
-             ("The same syncignore.txt and syncpoint.dat file logic applies and alows you to control the sync (please check Jari documentation)."),
-             (""),
-             ("NextSync source code can be found here: https://github.com/jarikomppa/specnext/tree/master/sync"),
-             (""),
-             ("If you run in any type of issue using the NextSync integration please run first the Jari command line version to see if it works as expected."),
-             (""),
-             ("OpenAL sound engine (on Windows)"),
-             ("--------------------------------"),
-             ("OpenAL library is required on Windows for CSpect to play sound, you may download it here: https://openal.org/"),
-             (""),
-             ("Mono (on Linux & MacOS Only)"),
-             ("-------"),
-             ("You will also need to install manualy mono-complete package for example using: sudo apt-get install mono-complete"),
-             (""),
-             ("Third-Party Content Sources (GetIt / ZXDB / zxArt):"),
-             ("----------------------------------------------------"),
-             ("zx-next-unite integrates three external databases to let you browse and download"),
-             ("Spectrum-related software and artwork directly from within the application."),
-             ("The application consumes their public APIs — it does not host, mirror, or"),
-             ("redistribute any of the files itself."),
-             (""),
-             ("GetIt:"),
-             ("  GetIt is a community-maintained archive of ZX Spectrum Next software."),
-             ("  API base URL : https://www.specnext.com/getit/"),
-             ("  The application queries the GetIt API to list and search files, then"),
-             ("  downloads them directly from the URLs returned by that API."),
-             (""),
-             ("ZXDB:"),
-             ("  ZXDB is an open-source database of ZX Spectrum and related software,"),
-             ("  maintained by the community at https://github.com/zxdb/ZXDB ."),
-             (f"  API base URL : {ZXDB_BASE_URL}"),
-             ("  The application queries the ZXDB REST API for titles, releases, screenshots"),
-             ("  and inlays, then downloads files directly from the URLs returned by that API."),
-             (""),
-             ("zxArt:"),
-             ("  zxArt (https://zxart.ee) is a gallery and archive dedicated to ZX Spectrum"),
-             ("  visual art, music, and productions."),
-             (f"  API base URL : {ZXART_BASE_URL}"),
-             ("  The application sends requests to the zxArt API to search productions and"),
-             ("  pictures, retrieve metadata and preview images, and download productions"),
-             ("  directly from the URLs returned by that API."),
-             (""),
-             ("Legal disclaimer:"),
-             ("  The author of zx-next-unite does NOT distribute any files, ROMs, games,"),
-             ("  demos, graphics, music, or any other content obtained through these APIs."),
-             ("  All content is served exclusively by the respective third-party services."),
-             ("  This application and author do not control third-party content."),
-             ("  It is the sole responsibility of the end user to ensure that any content"),
-             ("  they download or use through this application complies with the applicable"),
-             ("  copyright, licensing, and legal requirements in their jurisdiction."),
-             ("  If in doubt, consult the terms of service of the relevant platform and"),
-             ("  seek appropriate legal advice before downloading or using any content."),
-             ("  For inquiries you may reach out to me on my github page: https://github.com/jclauzel/ZX-Next-Unite"),
-             (""),
-             ("Enjoy!"),
-             ("")
-            )
 
 # Build the disclaimer text once from INIT_HELP (the "Legal disclaimer:" block)
 def _build_disclaimer_text():
@@ -807,228 +466,12 @@ def _make_disclaimer_ticker(parent):
     return lbl, timer
 
 
-CONFIG_FILE_SETTINGS = (SETTING_HDDFILE, SETTING_EXPLORERPATH, SETTING_SCREENSIZE, SETTING_SOUND, SETTING_VSYNC, SETTING_HERTZ, SETTING_JOYSTICK, SETTING_CSPECT, SETTING_CUSTOM, SETTING_ESC, SETTING_NEXTSYNC_EXPLORERPATH, SETTING_NEXTSYNC_SYNCONCE,
-SETTING_NEXTSYNC_ALWAYSSYNC, SETTING_NEXTSYNC_SLOWTRANSFER, SETTING_DEFAULT_TAB_WHEN_OPENING, SETTING_WARN_IMAGE_NEARLY_FULL, SETTING_NO_PROMPT_ON_DELETION, SETTING_COLOR_UP_DIRECTORY, SETTING_COLOR_DIR_NAME, SETTING_COLOR_DIR_TYPE, SETTING_COLOR_FILE_NAME,
-SETTING_COLOR_FILE_EXT, SETTING_COLOR_FILE_SIZE, SETTING_IMAGE_HISTORY, SETTING_ZXDB_LAST_MODE, SETTING_ZXDB_LAST_QUERY, SETTING_CONTENT_DISCLAIMER_AGREED, SETTING_BG_OPACITY, SETTING_AVAIL_CHECK, SETTING_MULTI_SEARCH, SETTING_SEARCH_AUTOCOMPLETE, SETTING_GALLERY_ANIM_MODE,
-SETTING_GALLERY_ROWS_PER_PAGE, SETTING_GALLERY_COLS, SETTING_GALLERY_IMG_SIZE, SETTING_GETIT_VIEW_MODE, SETTING_ZXDB_VIEW_MODE,
-SETTING_ZXART_VIEW_MODE, SETTING_ZXART_LANGUAGE, SETTING_FAVORITES, SETTING_FAVORITES_VIEW_MODE,
-SETTING_ALLINONE_VIEW_MODE, SETTING_BG_IMAGE, SETTING_CRASH_LOG_ENABLED)
-
-IMAGE_BUTTONS_SIZE = 190
-DISK_ARROWS_BUTTONS_SIZE = 30
-
-CSPECT_SCREEN_SIZES = (("Screen Size X1", "-w1"),("Screen Size X2", "-w2"),("Screen Size X3", "-w3"), ("Screen Size X4", "-w4"), ("Fullscreen", "-fullscreen"))
-CSPECT_SOUND = (("Sound On", ""),("Sound Off", "-sound"))
-CSPECT_SCREEN_SYNC = (("VSync On", "-vsync"),("VSync Off", ""))
-CSPECT_JOYSTICK = (("Joystick On", "-vsync"),("Joystick Off", ""))
-CSPECT_FREQUENCY = (("50Hz", ""),("60Hz", "-60"))
-CSPECT_BASE_ARGUMENTS = "-basickeys -zxnext -nextrom"
-
-FONT_GREEN = QColor(0, 255, 0)
-FONT_BLUE = QColor(0, 0, 255)
-FONT_RED = QColor(255, 0, 0)
-
-def qcolor_to_hex(color: QColor) -> str:
-    """Return a lowercase #rrggbb hex string for the given QColor."""
-    return color.name().lower()
-
-def hex_to_qcolor(hex_str: str) -> QColor:
-    """Return a QColor from a #rrggbb hex string, falling back to white on error."""
-    color = QColor(hex_str)
-    return color if color.isValid() else QColor(255, 255, 255)
-
-UP_DIRECTORY = "[Up Directory..]"
-DIRECTORY_CREATION_NOT_ALLOWED_CHARACTERS = ('"', '<', '>', ':', '\\', '/', '|', '?', '*', '!', '(',')', '.', "'", '$', '@')
-HDFMONKEY_EXECUTABLE = "hdfmonkey"
-FILTER_LABEL_TEXT = "Filter: "
-FILTER_TEXT_WIDTH = 320
-
 
 assert sys.version_info >= (3, 6) # We need 3.6 for f"" strings.
 
 # Configure logging
 logging.basicConfig(level=logging.INFO, format='%(asctime)s - %(levelname)s - %(message)s')
 
-class DotDotFirstProxyModel(QSortFilterProxyModel):
-    """Proxy model that always keeps the '..' parent directory entry at the top."""
-    def lessThan(self, left, right):
-        left_name = self.sourceModel().fileName(left)
-        right_name = self.sourceModel().fileName(right)
-        if left_name == "..":
-            return True
-        if right_name == "..":
-            return False
-        return super().lessThan(left, right)
-
-    def filterAcceptsRow(self, source_row, source_parent):
-        source_model = self.sourceModel()
-        index = source_model.index(source_row, 0, source_parent)
-        # Always show the parent-directory entry
-        if source_model.fileName(index) == "..":
-            return True
-        pattern = self.filterRegularExpression().pattern()
-        if not pattern:
-            return True
-        name = source_model.fileName(index)
-        return pattern.lower() in name.lower()
-
-class WorkerSignals(QObject):
-
-    finished = Signal()
-    error = Signal(tuple)
-    result = Signal(object)
-    progress = Signal(int)
-
-
-class NextSyncSignals(QObject):
-    """Signals used to marshal nextsync progress back to the main thread."""
-    progress = Signal(int)   # 0-100 per-file progress
-    status   = Signal(str)   # single-line status message
-    finished = Signal()      # emitted when the job thread exits
-    cancelled = Signal()     # emitted when job stopped due to cancel request
-
-
-class HdfTaskSignals(QObject):
-    """Signals for background hdfmonkey task workers."""
-    progress  = Signal(int)   # 0-100
-    status    = Signal(str)   # "action line\nfilename line"
-    finished  = Signal()
-    error     = Signal(str)   # human-readable error message
-    cancelled = Signal()      # emitted when the worker stopped early due to cancel
-
-
-class HdfTaskWorker(QRunnable):
-    """Generic QRunnable that runs a callable on the thread pool.
-    The callable receives (signals, cancel_event, *args, **kwargs).
-    Call worker.cancel() from the UI thread to request early termination."""
-
-    def __init__(self, fn, *args, **kwargs):
-        super().__init__()
-        self.fn           = fn
-        self.args         = args
-        self.kwargs       = kwargs
-        self.signals      = HdfTaskSignals()
-        self.cancel_event = threading.Event()
-        self.setAutoDelete(True)
-
-    def cancel(self):
-        self.cancel_event.set()
-
-    @Slot()
-    def run(self):
-        try:
-            self.fn(self.signals, self.cancel_event, *self.args, **self.kwargs)
-        except Exception as exc:
-            self.signals.error.emit(str(exc))
-        finally:
-            if self.cancel_event.is_set():
-                self.signals.cancelled.emit()
-            self.signals.finished.emit()
-
-
-class HdfProgressDialog(QDialog):
-    """Modal progress dialog with live status, progress bar, spinner, and Cancel button."""
-
-    cancel_requested = Signal()
-
-    def __init__(self, title, parent=None):
-        super().__init__(parent)
-        self.setWindowTitle(title)
-        self.setModal(True)
-        self.setMinimumWidth(540)
-        self.setWindowFlags(self.windowFlags() & ~Qt.WindowCloseButtonHint)
-
-        layout = QVBoxLayout(self)
-        layout.setSpacing(8)
-
-        # Spinner + action label on one row
-        action_row = QHBoxLayout()
-        self._spinner_label = QLabel("")
-        self._spinner_label.setFixedWidth(22)
-        action_row.addWidget(self._spinner_label)
-        self._action_label = QLabel("Starting\u2026")
-        self._action_label.setWordWrap(True)
-        action_row.addWidget(self._action_label, 1)
-        layout.addLayout(action_row)
-
-        # Current filename (smaller, muted)
-        self._file_label = QLabel("")
-        self._file_label.setWordWrap(True)
-        _font = self._file_label.font()
-        _ps = _font.pointSize()
-        if _ps <= 0:
-            _ps = max(QFontInfo(_font).pointSize(), 9)
-        _font.setPointSize(max(_ps - 1, 8))
-        self._file_label.setFont(_font)
-        layout.addWidget(self._file_label)
-
-        # Progress bar
-        self._bar = QProgressBar()
-        self._bar.setRange(0, 100)
-        self._bar.setValue(0)
-        self._bar.setTextVisible(True)
-        layout.addWidget(self._bar)
-
-        # Cancel button
-        btn_row = QHBoxLayout()
-        btn_row.addStretch()
-        self._cancel_btn = QPushButton("Cancel")
-        self._cancel_btn.setFixedWidth(90)
-        self._cancel_btn.clicked.connect(self._on_cancel_clicked)
-        btn_row.addWidget(self._cancel_btn)
-        layout.addLayout(btn_row)
-
-        self._cancelled = False
-        self._spinner_frames = ["\u25f4", "\u25f7", "\u25f6", "\u25f5"]
-        self._spinner_idx    = 0
-
-        self._anim_timer = QTimer(self)
-        self._anim_timer.setInterval(120)
-        self._anim_timer.timeout.connect(self._tick_spinner)
-        self._anim_timer.start()
-
-    # ------------------------------------------------------------------
-    @Slot()
-    def _on_cancel_clicked(self):
-        self._cancelled = True
-        self._cancel_btn.setEnabled(False)
-        self._action_label.setText("Cancelling\u2026")
-        self._file_label.setText("")
-        self.cancel_requested.emit()
-
-    @Slot()
-    def _tick_spinner(self):
-        self._spinner_idx = (self._spinner_idx + 1) % len(self._spinner_frames)
-        self._spinner_label.setText(self._spinner_frames[self._spinner_idx])
-
-    @Slot(int)
-    def set_progress(self, value: int):
-        """value == -1 activates the indeterminate (busy) marquee animation."""
-        if value < 0:
-            self._bar.setRange(0, 0)   # Qt marquee mode
-        else:
-            if self._bar.maximum() == 0:
-                self._bar.setRange(0, 100)
-            self._bar.setValue(value)
-
-    @Slot(str)
-    def set_status(self, text: str):
-        """Expects 'Action description\nFilename or detail'."""
-        if self._cancelled:
-            return
-        lines = text.split("\n", 1)
-        self._action_label.setText(lines[0])
-        self._file_label.setText(lines[1] if len(lines) > 1 else "")
-
-    @Slot()
-    def mark_cancelled(self):
-        """Called when the worker confirms it stopped early."""
-        self._action_label.setText("Cancelled.")
-        self._file_label.setText("")
-
-    def closeEvent(self, event):
-        self._anim_timer.stop()
-        super().closeEvent(event)
 
 
 # ---------------------------------------------------------------------------
@@ -2233,24 +1676,6 @@ def zxart_parse_prod_list(response: dict) -> tuple:
 # Selected zxART API language ("eng" | "pol" | "spa").  Mutated by the
 # language combo in the zxArt pane and persisted to the cfg file.  All
 # zxART HTTP request builders use _zxart_lang() to honour this value.
-_zxart_current_language: str = DEFAULT_ZXART_LANGUAGE
-
-
-def _zxart_lang() -> str:
-    """Return the currently-selected zxART API language code."""
-    lang = (_zxart_current_language or DEFAULT_ZXART_LANGUAGE).strip().lower()
-    if lang not in ("eng", "pol", "spa"):
-        lang = DEFAULT_ZXART_LANGUAGE
-    return lang
-
-
-def _zxart_set_language(code: str) -> None:
-    """Update the global zxART API language."""
-    global _zxart_current_language
-    code = (code or DEFAULT_ZXART_LANGUAGE).strip().lower()
-    if code not in ("eng", "pol", "spa"):
-        code = DEFAULT_ZXART_LANGUAGE
-    _zxart_current_language = code
 
 
 
@@ -2517,2455 +1942,10 @@ def getit_run_in_thread(fn, on_result, on_error):
     return t
 
 
-# ---------------------------------------------------------------------------
-# Gallery tag extraction (zxArt / ZXDB / GetIt agnostic helper)
-# ---------------------------------------------------------------------------
-#
-# zxART exposes a small set of short tag-like descriptors on each gallery
-# thumbnail (visible on https://zxart.ee/eng/mainpage/) such as:
-#   "GS"     – General Sound (sound hardware)
-#   "AY/YM"  – AY-3-8910 / YM2149 sound chip
-#   "Tape"   – tape-based release
-#   "TR-DOS" – TR-DOS (5.25"/3.5" disc)
-#   "48"     – Spectrum 48K
-#   "128"    – Spectrum 128K / +2
-#   "Pent"   – Pentagon / Pent128
-#   "KS8"    – Kondratiev Sprinter / KS8
-#   "TS"     – ZX-Evolution TS-Conf basic
-#   "TS-Conf"– TS-Conf advanced
-# The API does not return these as a single field; we derive them from
-# release formats, hardware requirements, and prod-level flags.
 
-# Base map: every extension from ZXFMT_PLACEHOLDER_LABELS gets its own label
-# so the overlay always shows something for any known format.  The grouped
-# overrides below replace generic labels with friendlier cluster names (e.g.
-# several tape formats all collapse to "Tape").
-_ZXART_RELEASE_FORMAT_TAGS: dict = {}
 
-# Populated lazily once ZXFMT_PLACEHOLDER_LABELS is defined (see below).
-def _build_release_format_tags() -> dict:
-    d = {ext: label for ext, label in ZXFMT_PLACEHOLDER_LABELS}
-    # Grouped / friendlier overrides
-    for _ext in ("tap", "tzx", "pzx", "cdt", "csw", "voc", "wav", "mp3", "ogg", "flac"):
-        if _ext in d:
-            d[_ext] = "Tape"
-    for _ext in ("trd", "scl", "fdi", "td0", "opd", "opu"):
-        if _ext in d:
-            d[_ext] = "TR-DOS"
-    for _ext in ("dsk", "mgt"):
-        if _ext in d:
-            d[_ext] = "+3 Disk"
-    for _ext in ("img", "hdf"):
-        if _ext in d:
-            d[_ext] = "HDD"
-    for _ext in ("sna", "z80", "szx", "sp", "zx", "slt"):
-        if _ext in d:
-            d[_ext] = "Snapshot"
-    for _ext in ("zip", "7z", "rar", "tar", "gz", "xz"):
-        if _ext in d:
-            d[_ext] = "Archive"
-    for _ext in ("bas", "asm", "z80s", "c", "h", "pas"):
-        if _ext in d:
-            d[_ext] = "Source"
-    for _ext in ("pdf", "txt", "doc", "docx", "rtf", "htm", "html", "nfo", "diz", "md"):
-        if _ext in d:
-            d[_ext] = "Docs"
-    return d
 
-_ZXART_HARDWARE_TAGS = {
-    # Sound / extension hardware
-    "generalsound":   "GS",
-    "gs":             "GS",
-    "ay":             "AY/YM",
-    "ay-3-8912":      "AY/YM",
-    "ym":             "AY/YM",
-    "beeper":         "Beeper",
-    "covox":          "Covox",
-    "soundrive":      "Soundrive",
-    # Storage / disc systems
-    "trdos":          "TR-DOS",
-    "tr-dos":         "TR-DOS",
-    "dosbase":        "TR-DOS",
-    # CPU / machine class (zxArt uses zx48/zx128/zx, etc.)
-    "zx":             "48",
-    "zx48":           "48",
-    "zxspectrum48":   "48",
-    "zx128":          "128",
-    "zxspectrum128":  "128",
-    "pentagon":       "Pent",
-    "pent":           "Pent",
-    "pent128":        "Pent128",
-    "pentagon128":    "Pent128",
-    "pent512":        "Pent512",
-    "pentagon512":    "Pent512",
-    "pent1024":       "Pent1024",
-    "pentagon1024":   "Pent1024",
-    "scorpion":       "Scorpion",
-    "scorpion256":    "Scorpion256",
-    "atm":            "ATM",
-    "atm2":           "ATM2",
-    "atm3":           "ATM3",
-    "ts-conf":        "TS-Conf",
-    "tsconf":         "TS-Conf",
-    "evolution":      "TS",
-    "zxevolution":    "TS",
-    "ks8":            "KS8",
-    "sprinter":       "KS8",
-    "zxnext":         "Next",
-    "next":           "Next",
-}
 
-
-def _zxart_tag_for_format(fmt: str) -> str:
-    if not fmt:
-        return ""
-    f = str(fmt).strip().lower().lstrip(".")
-    return _ZXART_RELEASE_FORMAT_TAGS.get(f, "")
-
-
-def _zxart_tag_for_hardware(name) -> str:
-    if not name:
-        return ""
-    n = str(name).strip().lower().replace(" ", "").replace("_", "")
-    return _ZXART_HARDWARE_TAGS.get(n, "")
-
-
-def _gallery_extract_tags(entry: dict) -> list:
-    """Return a de-duplicated list of short overlay tags for a gallery entry.
-
-    For zxArt entries we look at the production's release formats and
-    hardware-required list. For pictures we surface the picture type (e.g.
-    "standard", "border", "multicolor"). For non-zxArt entries we fall back
-    to the entry's `machine` and `genre` keys when they look tag-like.
-    """
-    if not isinstance(entry, dict):
-        return []
-
-    tags = []
-    seen = set()
-
-    def _add(t):
-        t = (t or "").strip()
-        if not t:
-            return
-        k = t.lower()
-        if k in seen:
-            return
-        seen.add(k)
-        tags.append(t)
-
-    src  = entry.get("_source") or {}
-    kind = (entry.get("_kind") or "").lower()
-
-    if kind == "zxart_picture":
-        ptype = src.get("type") or ""
-        if ptype:
-            _add(str(ptype))
-        for t in (src.get("tags") or [])[:3]:
-            _add(str(t))
-    else:
-        # Prefer release-level info (most accurate); fall back to prod fields.
-        for rel in (src.get("releases") or []):
-            if not isinstance(rel, dict):
-                continue
-            rf = rel.get("releaseFormat")
-            if isinstance(rf, list):
-                for f in rf:
-                    _add(_zxart_tag_for_format(f))
-            else:
-                _add(_zxart_tag_for_format(rf or ""))
-            for hw in (rel.get("hardwareRequired") or []):
-                _add(_zxart_tag_for_hardware(hw))
-        for fmt in (src.get("releaseFormats") or []):
-            _add(_zxart_tag_for_format(fmt))
-        for hw in (src.get("hardwareRequired") or []):
-            _add(_zxart_tag_for_hardware(hw))
-        # Some prod payloads expose machine names in plain text.
-        machine = entry.get("machine") or ""
-        if machine:
-            for piece in str(machine).replace(",", " ").split():
-                _add(_zxart_tag_for_hardware(piece))
-
-    # Cap the badge count so the overlay stays readable.
-    return tags[:5]
-
-
-# ---------------------------------------------------------------------------
-# ZX Spectrum file-format helpers (used for gallery placeholder thumbnails
-# when a content entry has no actual picture available).  Reference list:
-#   https://worldofspectrum.org/faq/reference/formats.htm
-# ---------------------------------------------------------------------------
-
-# Recognised image-bearing extensions (real screenshots/cover scans).
-ZXFMT_IMAGE_EXTS = (
-    ".png", ".jpg", ".jpeg", ".gif", ".bmp", ".webp",
-    ".scr",                 # Spectrum native screen$
-    ".mlt",                 # MULTITECH ULAplus 8x1 attribute screen
-    ".ifl", ".bsc", ".mc",  # 8x1, BiFROST, multicolour screens
-    ".nxi",                 # ZX Spectrum Next layer-2 raw screen
-    ".slr", ".sli",         # SamRam Spectrum / scanline images
-    ".chr$",                # Char-mapped images
-)
-
-# Format -> short tag shown in big letters in the placeholder thumbnail.
-# Keep ordering meaningful (specific before generic).  Bucketed by category.
-ZXFMT_PLACEHOLDER_LABELS = (
-    # Snapshots
-    ("sna",  "SNA"),
-    ("z80",  "Z80"),
-    ("szx",  "SZX"),
-    ("sp",   "SP"),
-    ("zx",   "ZX"),
-    ("slt",  "SLT"),
-    # Tape images
-    ("tap",  "TAP"),
-    ("tzx",  "TZX"),
-    ("pzx",  "PZX"),
-    ("cdt",  "CDT"),
-    ("csw",  "CSW"),
-    ("voc",  "VOC"),
-    ("wav",  "WAV"),
-    ("mp3",  "MP3"),
-    ("ogg",  "OGG"),
-    ("flac", "FLAC"),
-    # Disk images
-    ("dsk",  "DSK"),
-    ("trd",  "TRD"),
-    ("scl",  "SCL"),
-    ("fdi",  "FDI"),
-    ("td0",  "TD0"),
-    ("img",  "IMG"),
-    ("mgt",  "MGT"),
-    ("opd",  "OPD"),
-    ("opu",  "OPU"),
-    ("hdf",  "HDF"),
-    ("$b",   "$B"),
-    ("$c",   "$C"),
-    # Cartridges / ROM
-    ("rom",  "ROM"),
-    ("dck",  "DCK"),
-    # Pokes / cheats
-    ("pok",  "POK"),
-    # Source / code
-    ("bas",  "BAS"),
-    ("asm",  "ASM"),
-    ("z80s", "ASM"),
-    ("c",    "C"),
-    ("h",    "H"),
-    ("pas",  "PAS"),
-    # Documents
-    ("pdf",  "PDF"),
-    ("txt",  "TXT"),
-    ("doc",  "DOC"),
-    ("docx", "DOCX"),
-    ("rtf",  "RTF"),
-    ("htm",  "HTM"),
-    ("html", "HTML"),
-    ("nfo",  "NFO"),
-    ("diz",  "DIZ"),
-    ("md",   "MD"),
-    # Archives
-    ("zip",  "ZIP"),
-    ("7z",   "7Z"),
-    ("rar",  "RAR"),
-    ("tar",  "TAR"),
-    ("gz",   "GZ"),
-    ("xz",   "XZ"),
-    # Misc / fallback
-    ("xml",  "XML"),
-    ("json", "JSON"),
-)
-
-# Now that ZXFMT_PLACEHOLDER_LABELS is defined, build the release-format tag
-# map used by _zxart_tag_for_format() (defined earlier in the file).
-_ZXART_RELEASE_FORMAT_TAGS.update(_build_release_format_tags())
-
-
-def zxfmt_split_ext(name: str) -> str:
-    """Return the lower-case extension (without the dot) of *name*, handling
-    composite/uncommon Spectrum extensions like ``.tap.zip`` and ``$b``."""
-    if not name:
-        return ""
-    n = name.lower().strip()
-    # Strip query strings / fragments if a URL was passed in.
-    for sep in ("?", "#"):
-        if sep in n:
-            n = n.split(sep, 1)[0]
-    n = os.path.basename(n)
-    # Composite ``foo.tap.zip`` -> use the inner format as the primary hint.
-    parts = n.split(".")
-    if len(parts) >= 3 and parts[-1] in ("zip", "gz", "xz", "7z", "rar"):
-        inner = parts[-2]
-        if inner:
-            return inner
-    if "." in n:
-        return n.rsplit(".", 1)[-1]
-    # Files like ``foo$b`` (BASIC) or ``foo$c`` (CODE) come from +D / DISCiPLE.
-    if "$" in n:
-        return "$" + n.rsplit("$", 1)[-1]
-    return ""
-
-
-def zxfmt_is_image_name(name: str) -> bool:
-    """Return True if *name* points at an actual picture (not a generic
-    document or binary)."""
-    if not name:
-        return False
-    n = name.lower()
-    for sep in ("?", "#"):
-        if sep in n:
-            n = n.split(sep, 1)[0]
-    return any(n.endswith(ext) for ext in ZXFMT_IMAGE_EXTS)
-
-
-def zxfmt_label_for_name(name: str) -> str:
-    """Return a short upper-case label to render in a placeholder thumbnail
-    for *name* (e.g. ``"TAP"``, ``"PDF"``, ``"POK"``).  Falls back to the
-    extension itself, then to ``"FILE"``."""
-    ext = zxfmt_split_ext(name)
-    if not ext:
-        return "FILE"
-    for needle, label in ZXFMT_PLACEHOLDER_LABELS:
-        if ext == needle:
-            return label
-    return ext.upper()[:6]
-
-
-def zxfmt_make_placeholder_pixmap(label: str, subtitle: str = "",
-                                  width: int = 320) -> QPixmap:
-    """Render a 4:3 placeholder QPixmap with *label* in big letters and an
-    optional *subtitle* (typically the file name) below.  Used by the
-    gallery cells when no real screenshot is available so the user can still
-    tell at a glance what kind of content the entry holds."""
-    w = max(160, int(width))
-    h = int(w * 3 / 4)
-    pm = QPixmap(w, h)
-    pm.fill(QColor("#101820"))
-    p = QPainter(pm)
-    try:
-        p.setRenderHint(QPainter.Antialiasing, True)
-        p.setRenderHint(QPainter.TextAntialiasing, True)
-        # Frame
-        p.setPen(QColor("#3a4a5a"))
-        p.drawRect(2, 2, w - 5, h - 5)
-        # Big label
-        big = QFont()
-        big.setBold(True)
-        # Scale font size to roughly fit the label width.
-        target_w = int(w * 0.7)
-        size = max(18, int(h * 0.42))
-        big.setPointSize(size)
-        p.setFont(big)
-        # Reduce until the label fits.
-        for _ in range(12):
-            fm = p.fontMetrics()
-            if fm.horizontalAdvance(label) <= target_w:
-                break
-            size = max(10, size - 2)
-            big.setPointSize(size)
-            p.setFont(big)
-        p.setPen(QColor("#ffd24a"))
-        fm = p.fontMetrics()
-        label_h = fm.height()
-        label_y = int(h / 2) - 4
-        p.drawText(QRect(0, label_y - label_h, w, label_h),
-                   Qt.AlignCenter, label)
-        # Subtitle (file name) underneath
-        if subtitle:
-            sub = QFont()
-            sub.setPointSize(max(8, int(h * 0.09)))
-            p.setFont(sub)
-            p.setPen(QColor("#cfd8dc"))
-            fm2 = p.fontMetrics()
-            sub_text = subtitle
-            avail = w - 16
-            if fm2.horizontalAdvance(sub_text) > avail:
-                sub_text = fm2.elidedText(sub_text, Qt.ElideMiddle, avail)
-            sub_y = label_y + 6
-            p.drawText(QRect(8, sub_y, w - 16, fm2.height() + 4),
-                       Qt.AlignHCenter | Qt.AlignTop, sub_text)
-    finally:
-        p.end()
-    return pm
-
-
-# ---------------------------------------------------------------------------
-# ZX Spectrum .scr screen decoder
-# ---------------------------------------------------------------------------
-# The standard Spectrum screen file is a raw memory dump of the display area
-# at address 0x4000..0x5AFF, i.e. exactly 6912 bytes:
-#   * Bytes 0..6143   : 6144 bytes of bitmap (256 x 192 px, 1 bit per pixel).
-#   * Bytes 6144..6911:  768 bytes of attribute cells (32 x 24, 8x8 pixels each).
-#
-# Bitmap row addressing follows the well-known Spectrum interleave: for a
-# pixel row y in 0..191, the byte offset of the leftmost byte of that row is
-#       offset = ((y & 0b11000000) << 5)   # third = y / 64
-#              | ((y & 0b00000111) << 8)   # pixel row within char row
-#              | ((y & 0b00111000) << 2)   # char row within third
-# (each "third" is 2048 bytes; within a third, eight 256-byte pixel-row
-# planes are interleaved with the eight 32-byte text rows.)
-#
-# Attribute byte layout (per 8x8 cell):
-#   bit  7  : FLASH
-#   bit  6  : BRIGHT
-#   bits 5-3: PAPER colour (0..7)
-#   bits 2-0: INK   colour (0..7)
-#
-# The 8 standard Spectrum colours, as 24-bit RGB.  "Bright" uses 0xFF for the
-# non-zero components, "normal" uses 0xCD (a widely used approximation of the
-# real CRT brightness used by emulators and converters).
-ZXSCR_PALETTE_NORMAL = (
-    (0x00, 0x00, 0x00),  # 0 black
-    (0x00, 0x00, 0xCD),  # 1 blue
-    (0xCD, 0x00, 0x00),  # 2 red
-    (0xCD, 0x00, 0xCD),  # 3 magenta
-    (0x00, 0xCD, 0x00),  # 4 green
-    (0x00, 0xCD, 0xCD),  # 5 cyan
-    (0xCD, 0xCD, 0x00),  # 6 yellow
-    (0xCD, 0xCD, 0xCD),  # 7 white
-)
-ZXSCR_PALETTE_BRIGHT = (
-    (0x00, 0x00, 0x00),
-    (0x00, 0x00, 0xFF),
-    (0xFF, 0x00, 0x00),
-    (0xFF, 0x00, 0xFF),
-    (0x00, 0xFF, 0x00),
-    (0x00, 0xFF, 0xFF),
-    (0xFF, 0xFF, 0x00),
-    (0xFF, 0xFF, 0xFF),
-)
-ZXSCR_BYTES = 6912
-
-
-def zxscr_is_screen_bytes(data) -> bool:
-    """Return True if *data* looks like a 6912-byte Spectrum screen dump."""
-    return isinstance(data, (bytes, bytearray)) and len(data) == ZXSCR_BYTES
-
-
-def _zxscr_pixel_row_offset(y: int) -> int:
-    """Offset of the leftmost byte of pixel row *y* in the bitmap plane."""
-    return ((y & 0xC0) << 5) | ((y & 0x07) << 8) | ((y & 0x38) << 2)
-
-
-class ZxSpectrumScreen:
-    """Decoder for a standard 6912-byte ZX Spectrum SCREEN$ file.
-
-    This is an original Python implementation written from the public
-    description of the Spectrum display format; it does not derive from
-    any third-party converter.  ``flash_phase`` selects which half of the
-    FLASH cycle to render (False = INK/PAPER as stored, True = swapped).
-    """
-
-    def __init__(self, data: bytes):
-        if not zxscr_is_screen_bytes(data):
-            raise ValueError("not a 6912-byte SCR screen")
-        self._data = bytes(data)
-
-    @property
-    def width(self) -> int:
-        return 256
-
-    @property
-    def height(self) -> int:
-        return 192
-
-    def to_qimage(self, flash_phase: bool = False) -> QImage:
-        """Render the screen to a 256x192 ``QImage`` (RGB32)."""
-        img = QImage(256, 192, QImage.Format_RGB32)
-        data = self._data
-        attrs = data  # alias for clarity; offset 6144 onwards
-        for y in range(192):
-            row_off = _zxscr_pixel_row_offset(y)
-            attr_row = (y >> 3) * 32 + 6144
-            char_y = y >> 3  # not used after row_off but kept for clarity
-            for cx in range(32):
-                byte = data[row_off + cx]
-                attr = attrs[attr_row + cx]
-                ink   = attr & 0x07
-                paper = (attr >> 3) & 0x07
-                bright = bool(attr & 0x40)
-                flash  = bool(attr & 0x80)
-                if flash and flash_phase:
-                    ink, paper = paper, ink
-                pal = ZXSCR_PALETTE_BRIGHT if bright else ZXSCR_PALETTE_NORMAL
-                ink_r, ink_g, ink_b = pal[ink]
-                pap_r, pap_g, pap_b = pal[paper]
-                ink_px = (0xFF << 24) | (ink_r << 16) | (ink_g << 8) | ink_b
-                pap_px = (0xFF << 24) | (pap_r << 16) | (pap_g << 8) | pap_b
-                x = cx * 8
-                # MSB is the leftmost pixel of the byte.
-                for bit in range(8):
-                    on = (byte >> (7 - bit)) & 1
-                    img.setPixel(x + bit, y, ink_px if on else pap_px)
-        return img
-
-
-def zxscr_url_is_scr(url) -> bool:
-    """Return True if *url* (string) points at a .scr screen file."""
-    if not isinstance(url, str):
-        return False
-    n = url.lower()
-    for sep in ("?", "#"):
-        if sep in n:
-            n = n.split(sep, 1)[0]
-    return n.endswith(".scr")
-
-
-# In-memory cache: url/base_name -> QPixmap (avoids repeated decoding)
-# Guarded by a lock because it is populated from background worker threads
-# (SCR byte conversion) and read from the GUI thread (extra-fetch callbacks).
-_ZXSCR_PIXMAP_CACHE_LOCK = threading.RLock()
-_ZXSCR_PIXMAP_CACHE: dict = {}
-
-
-def _zxscr_basename_for_url(url: str) -> str:
-    """Derive a cache key (no extension) from *url*."""
-    try:
-        from urllib.parse import urlsplit, unquote
-        path = unquote(urlsplit(url).path)
-    except Exception:
-        path = url
-    name = os.path.basename(path) or "screen.scr"
-    stem, _ext = os.path.splitext(name)
-    safe = "".join(ch if (ch.isalnum() or ch in "-_.") else "_" for ch in stem)
-    return safe or "screen"
-
-
-def zxscr_convert_bytes_to_pixmap(data: bytes, base_name: str):
-    """Decode *data* (6912-byte SCR) and cache the QPixmap in memory.
-    Returns a QPixmap or None on failure."""
-    if not zxscr_is_screen_bytes(data):
-        return None
-    with _ZXSCR_PIXMAP_CACHE_LOCK:
-        cached = _ZXSCR_PIXMAP_CACHE.get(base_name)
-    if cached is not None:
-        return cached
-    try:
-        img = ZxSpectrumScreen(data).to_qimage()
-        pm = QPixmap.fromImage(img)
-        if not pm.isNull():
-            with _ZXSCR_PIXMAP_CACHE_LOCK:
-                _ZXSCR_PIXMAP_CACHE[base_name] = pm
-        return pm
-    except Exception:
-        return None
-
-
-def zxscr_try_pixmap_from_url_bytes(url, data):
-    """If *url* / *data* describe a Spectrum .scr screen, return a QPixmap
-    rendered from it (with caching).  Otherwise return ``None`` so the
-    caller can fall back to its normal image-loading path."""
-    if not zxscr_is_screen_bytes(data) and not zxscr_url_is_scr(url):
-        return None
-    if not zxscr_is_screen_bytes(data):
-        return None
-    base = _zxscr_basename_for_url(url) if isinstance(url, str) else "screen"
-    return zxscr_convert_bytes_to_pixmap(data, base)
-
-
-def zxfmt_pick_best_download(downloads):
-    """From a list of download dicts (each may have ``url``/``filename``/
-    ``format``/``type``), pick the entry most suitable for a placeholder
-    thumbnail label.  Preference order: tape > disk > snapshot > pok > docs
-    > anything else.  Returns ``(label, filename)`` or ``("FILE", "")`` if
-    *downloads* is empty / unusable."""
-    if not downloads:
-        return ("FILE", "")
-    priority = {
-        # tapes
-        "tap": 0, "tzx": 0, "pzx": 0, "cdt": 0, "csw": 1,
-        # disks
-        "trd": 2, "scl": 2, "dsk": 2, "fdi": 2, "td0": 2, "mgt": 2, "img": 2,
-        # snapshots
-        "sna": 3, "z80": 3, "szx": 3, "sp": 3, "slt": 3,
-        # cartridges
-        "rom": 4, "dck": 4,
-        # pokes
-        "pok": 5,
-        # docs
-        "pdf": 6, "txt": 6, "nfo": 6, "diz": 6,
-        # archives
-        "zip": 8, "7z": 8, "rar": 8, "tar": 8, "gz": 8, "xz": 8,
-    }
-    best = None
-    best_rank = 999
-    for d in downloads:
-        if not isinstance(d, dict):
-            continue
-        name = str(d.get("filename") or "")
-        url  = str(d.get("url") or "")
-        ref  = name or url
-        if not ref:
-            continue
-        ext  = zxfmt_split_ext(ref)
-        rank = priority.get(ext, 7)
-        if rank < best_rank:
-            best_rank = rank
-            best = (zxfmt_label_for_name(ref), name or os.path.basename(
-                urllib.parse.urlparse(url).path) if url else name)
-    if best is None:
-        # Fall back to the first usable entry.
-        for d in downloads:
-            if not isinstance(d, dict):
-                continue
-            name = str(d.get("filename") or "")
-            url  = str(d.get("url") or "")
-            ref  = name or url
-            if not ref:
-                continue
-            return (zxfmt_label_for_name(ref),
-                    name or os.path.basename(urllib.parse.urlparse(url).path))
-        return ("FILE", "")
-    return best
-
-
-def _build_tooltip_text(lines: list) -> str:
-    """Join `lines` into a tooltip string, truncating to MAX_ALT_TEXT_LINES.
-
-    If the total number of non-empty lines exceeds MAX_ALT_TEXT_LINES, the
-    list is cut at that limit and a trailing "..." line is appended.
-    """
-    non_empty = [ln for ln in lines if ln and str(ln).strip()]
-    if len(non_empty) > MAX_ALT_TEXT_LINES:
-        non_empty = non_empty[:MAX_ALT_TEXT_LINES] + ["..."]
-    return "\n".join(non_empty)
-
-
-class GalleryCell(QFrame):
-    """A picture-view tile: thumbnail + title + (hover-only) info line.
-
-    The cell is fully asynchronous:
-      * `thumb_fetch_cb(entry, set_pixmap, set_screenshots)` is invoked once
-        when the cell is shown. The callback is responsible for resolving the
-        main screenshot bytes off the UI thread and calling `set_pixmap(QPixmap)`
-        on the UI thread. If extra screenshots become available it should call
-        `set_screenshots([url, ...])` and the cell will then start cycling.
-      * `extra_fetch_cb(url, on_pixmap)` is invoked lazily when a new url is
-        about to be displayed and the pixmap is not yet cached.
-
-    Selection is communicated via the `clicked` signal which forwards the
-    entry dict. Animation honors the global animation mode ("hover"/"timer").
-    """
-
-    clicked     = Signal(object)
-    dbl_clicked = Signal(object)
-
-    _PLACEHOLDER_COLOR = QColor("#222")
-
-    def __init__(self, entry, anim_mode_getter,
-                 thumb_fetch_cb, extra_fetch_cb,
-                 title_text="", info_text="", tooltip_text="",
-                 context_menu_cb=None,
-                 tags=None, parent=None,
-                 is_favorite_cb=None, toggle_favorite_cb=None,
-                 source_label_getter=None,
-                 source_overlay_anchor="topleft"):
-        super().__init__(parent)
-        if tooltip_text:
-            self.setToolTip(tooltip_text)
-        self._entry = entry
-        self._anim_mode_getter = anim_mode_getter  # callable -> "hover"|"timer"
-        self._thumb_fetch_cb = thumb_fetch_cb
-        self._extra_fetch_cb = extra_fetch_cb
-        self._context_menu_cb = context_menu_cb
-        self._is_favorite_cb     = is_favorite_cb
-        self._toggle_favorite_cb = toggle_favorite_cb
-        self._source_label_getter = source_label_getter
-        self._source_overlay_anchor = source_overlay_anchor or "topleft"
-        self._screenshots = []        # list of URL strings (or dicts {"url": ...})
-        self._shot_cache  = {}        # url -> QPixmap
-        self._shot_index  = 0
-        self._hovered     = False
-        self._selected    = False
-        self._thumb_w     = 160       # last applied width hint
-        self._tags        = [str(t) for t in (tags or []) if t]
-        # Callable set by GalleryView when this cell ends up with no usable
-        # image (either an explicit empty set_screenshots, or every URL in
-        # the cycling list failed to load).  The view uses it to move the
-        # cell to the end of the grid so the user sees items with pictures
-        # first.  We call it at most once per cell.
-        self._no_image_cb = None
-        self._no_image_notified = False
-        # Callable(cell, QPixmap, url) set by GalleryView so a resolved real
-        # picture can be cached for instant redraw across an image-first
-        # re-sort.  Called whenever a non-placeholder pixmap is applied.
-        self._pixmap_ready_cb = None
-        self._last_applied_url = ""
-
-        self.setFrameShape(QFrame.StyledPanel)
-        self.setFrameShadow(QFrame.Plain)
-        self.setLineWidth(1)
-        self.setAttribute(Qt.WA_StyledBackground, True)
-        self.setMouseTracking(True)
-        self.setCursor(Qt.PointingHandCursor)
-        self._apply_style()
-
-        lay = QVBoxLayout(self)
-        lay.setContentsMargins(4, 4, 4, 4)
-        lay.setSpacing(2)
-
-        self._thumb_lbl = QLabel(self)
-        self._thumb_lbl.setAlignment(Qt.AlignCenter)
-        self._thumb_lbl.setMinimumHeight(60)
-        self._thumb_lbl.setStyleSheet("background-color: #222; color: #888;")
-        self._thumb_lbl.setText("…")
-        self._thumb_lbl.setSizePolicy(QSizePolicy.Expanding, QSizePolicy.Expanding)
-        lay.addWidget(self._thumb_lbl, 1)
-
-        # Overlay tag badges (top-right of the thumbnail). Floats above the
-        # pixmap via a child QLabel that we re-position in resizeEvent.
-        self._tag_overlay = QLabel(self._thumb_lbl)
-        self._tag_overlay.setAttribute(Qt.WA_TransparentForMouseEvents, True)
-        self._tag_overlay.setAlignment(Qt.AlignTop | Qt.AlignRight)
-        self._tag_overlay.setTextFormat(Qt.RichText)
-        self._tag_overlay.setStyleSheet("background: transparent;")
-        self._tag_overlay.setVisible(False)
-        self._refresh_tag_overlay()
-
-        # Heart (favorite) button overlay – bottom-right of the thumbnail.
-        self._heart_btn = QToolButton(self._thumb_lbl)
-        self._heart_btn.setCursor(Qt.PointingHandCursor)
-        self._heart_btn.setAutoRaise(True)
-        self._heart_btn.setFocusPolicy(Qt.NoFocus)
-        self._heart_btn.setText("♡")
-        self._heart_btn.setToolTip("Add to favorites")
-        self._heart_btn.setStyleSheet(
-            "QToolButton { color: #ff5577; background: rgba(0,0,0,140);"
-            " border: 1px solid rgba(255,255,255,40); border-radius: 10px;"
-            " padding: 0px 4px; font-size: 14pt; font-weight: bold; }"
-            "QToolButton:hover { background: rgba(0,0,0,200);"
-            " border-color: #ff5577; }"
-        )
-        self._heart_btn.setVisible(bool(self._toggle_favorite_cb))
-        self._heart_btn.clicked.connect(self._on_heart_clicked)
-
-        # Source-pane overlay badge (top-left of thumbnail).  Used by the
-        # Favorites gallery to show whether an item came from GetIt / ZXDB /
-        # zxArt.  Hidden by default for the regular galleries.
-        self._source_overlay = QLabel(self._thumb_lbl)
-        self._source_overlay.setAttribute(Qt.WA_TransparentForMouseEvents, True)
-        self._source_overlay.setAlignment(Qt.AlignTop | Qt.AlignLeft)
-        self._source_overlay.setTextFormat(Qt.RichText)
-        self._source_overlay.setStyleSheet("background: transparent;")
-        self._source_overlay.setVisible(False)
-        self._refresh_source_overlay()
-        self._refresh_heart()
-
-        self._title_lbl = QLabel(title_text or "", self)
-        self._title_lbl.setAlignment(Qt.AlignCenter)
-        self._title_lbl.setWordWrap(True)
-        self._title_lbl.setTextFormat(Qt.RichText)
-        f = self._title_lbl.font()
-        f.setBold(True)
-        self._title_lbl.setFont(f)
-        lay.addWidget(self._title_lbl, 0)
-
-        self._info_lbl = QLabel(info_text or "", self)
-        self._info_lbl.setAlignment(Qt.AlignCenter)
-        self._info_lbl.setWordWrap(True)
-        self._info_lbl.setStyleSheet("color: #aaa;")
-        self._info_lbl.setVisible(False)
-        lay.addWidget(self._info_lbl, 0)
-
-        self._timer = QTimer(self)
-        self._timer.setInterval(1500)
-        self._timer.timeout.connect(self._advance)
-
-        # Retry bookkeeping for the initial thumbnail fetch.  The per-gallery
-        # fetch callbacks run on background threads whose error path is a no-op,
-        # so a transient network/JSON failure leaves the cell stuck on the "…"
-        # placeholder forever (until the user opens the item full-screen, which
-        # kicks off a fresh fetch).  We retry the initial fetch a few times with
-        # back-off whenever the cell has neither loaded a real picture nor
-        # resolved to a typed placeholder.
-        self._loaded_ok          = False   # a non-placeholder pixmap was shown
-        self._placeholder_shown  = False   # a "no real picture" placeholder shown
-        self._fetch_attempts     = 0
-        self._max_fetch_attempts = 4
-        self._retry_timer = QTimer(self)
-        self._retry_timer.setSingleShot(True)
-        self._retry_timer.timeout.connect(self._retry_if_unloaded)
-
-        # Defer initial fetch until the cell is actually shown
-        QTimer.singleShot(0, self._kickoff_initial_fetch)
-
-    # ---- public API ----------------------------------------------------
-
-    def entry(self):
-        return self._entry
-
-    def set_selected(self, selected: bool):
-        if self._selected == selected:
-            return
-        self._selected = selected
-        self._apply_style()
-
-    def set_thumb_width(self, w: int):
-        """Called by the parent gallery view when the column width changes."""
-        w = max(40, int(w))
-        changed = abs(w - self._thumb_w) >= 4
-        self._thumb_w = w
-        # Always rescale the current pixmap. Even when the requested width is
-        # unchanged, the pixmap may have been left at the wrong size after
-        # returning from a fullscreen / in-pane viewer where the underlying
-        # QLabel briefly reported a much larger geometry.
-        cur = self._current_pixmap()
-        if cur is not None:
-            self._apply_pixmap(cur)
-        return changed
-
-    def set_tags(self, tags):
-        """Replace the overlay tag badges."""
-        self._tags = [str(t) for t in (tags or []) if t]
-        self._refresh_tag_overlay()
-
-    def set_info_text(self, text: str):
-        """Update the hover-only info line (author / date / category, …).
-
-        Called asynchronously by thumb-fetch callbacks once richer metadata
-        becomes available (e.g. GetIt detail records that include the date).
-        """
-        try:
-            self._info_lbl.setText(text or "")
-        except Exception:
-            pass
-
-    def _refresh_tag_overlay(self):
-        if not getattr(self, "_tag_overlay", None):
-            return
-        if not self._tags:
-            self._tag_overlay.setVisible(False)
-            self._tag_overlay.setText("")
-            return
-        # Build a row of HTML chips. Inline-block keeps them right-aligned and
-        # wrapping naturally on narrow tiles.
-        chip_css = (
-            "background:#1c3a52;color:#bfe6ff;border:1px solid #2f6f9a;"
-            "border-radius:3px;padding:1px 5px;margin:1px 2px;"
-            "font-size:9pt;font-weight:600;"
-        )
-        chips = "&nbsp;".join(
-            f"<span style='{chip_css}'>{t}</span>"
-            for t in self._tags
-        )
-        html = f"<div style='text-align:right;'>{chips}</div>"
-        self._tag_overlay.setText(html)
-        self._tag_overlay.setVisible(True)
-        self._position_tag_overlay()
-
-    def _position_tag_overlay(self):
-        ov = getattr(self, "_tag_overlay", None)
-        if ov is None or not ov.isVisible():
-            return
-        # Anchor to top-right with a small padding inside the thumb label.
-        pad = 4
-        max_w = max(40, self._thumb_lbl.width() - 2 * pad)
-        ov.adjustSize()
-        w = min(ov.width(), max_w)
-        h = ov.sizeHint().height()
-        ov.setGeometry(self._thumb_lbl.width() - w - pad, pad, w, h)
-        ov.raise_()
-
-    def _refresh_source_overlay(self):
-        ov = getattr(self, "_source_overlay", None)
-        if ov is None:
-            return
-        getter = self._source_label_getter
-        label = ""
-        if getter is not None:
-            try:
-                label = str(getter(self._entry) or "").strip()
-            except Exception:
-                label = ""
-        if not label:
-            ov.setVisible(False)
-            ov.setText("")
-            return
-        chip_css = (
-            "background:#3a1c52;color:#e8bfff;border:1px solid #6f2f9a;"
-            "border-radius:3px;padding:1px 6px;margin:1px 2px;"
-            "font-size:9pt;font-weight:600;"
-        )
-        ov.setText(f"<div style='text-align:left;'>"
-                   f"<span style='{chip_css}'>{label}</span></div>")
-        ov.setVisible(True)
-        self._position_source_overlay()
-
-    def _position_source_overlay(self):
-        ov = getattr(self, "_source_overlay", None)
-        if ov is None or not ov.isVisible():
-            return
-        pad = 4
-        ov.adjustSize()
-        w = min(ov.width(), max(40, self._thumb_lbl.width() - 2 * pad))
-        h = ov.sizeHint().height()
-        anchor = getattr(self, "_source_overlay_anchor", "topleft")
-        if anchor == "bottomleft":
-            x = pad
-            y = self._thumb_lbl.height() - h - pad
-        elif anchor == "bottomright":
-            # Sit to the left of the heart button (which is anchored bottom-
-            # right) if one is shown, otherwise hug the right edge.
-            heart = getattr(self, "_heart_btn", None)
-            heart_w = 0
-            if heart is not None and heart.isVisible():
-                heart_w = heart.sizeHint().width() + pad
-            x = max(pad, self._thumb_lbl.width() - w - pad - heart_w)
-            y = self._thumb_lbl.height() - h - pad
-        else:
-            x = pad
-            y = pad
-        ov.setGeometry(x, y, w, h)
-        ov.raise_()
-
-    def _refresh_heart(self):
-        btn = getattr(self, "_heart_btn", None)
-        if btn is None:
-            return
-        if not self._toggle_favorite_cb:
-            btn.setVisible(False)
-            return
-        is_fav = False
-        if self._is_favorite_cb is not None:
-            try:
-                is_fav = bool(self._is_favorite_cb(self._entry))
-            except Exception:
-                is_fav = False
-        btn.setText("♥" if is_fav else "♡")
-        btn.setToolTip("Remove from favorites" if is_fav else "Add to favorites")
-        btn.setVisible(True)
-        self._position_heart()
-
-    def _position_heart(self):
-        btn = getattr(self, "_heart_btn", None)
-        if btn is None or not btn.isVisible():
-            return
-        pad = 4
-        btn.adjustSize()
-        w = btn.sizeHint().width()
-        h = btn.sizeHint().height()
-        btn.setGeometry(self._thumb_lbl.width() - w - pad,
-                         self._thumb_lbl.height() - h - pad, w, h)
-        btn.raise_()
-
-    def _on_heart_clicked(self):
-        if not self._toggle_favorite_cb:
-            return
-        try:
-            self._toggle_favorite_cb(self._entry)
-        except Exception:
-            pass
-        self._refresh_heart()
-
-    def set_favorite_state_changed(self):
-        """External hook: re-query the favorite predicate (e.g. when another
-        view toggled the favorite state for this entry)."""
-        self._refresh_heart()
-
-    def _is_alive(self) -> bool:
-        """Return False if the underlying C++ QWidget has already been deleted
-        (which happens when GalleryView.populate() tears cells down while
-        background thumbnail fetches are still in flight)."""
-        try:
-            # Touching any property forces shiboken to validate the C++ object.
-            self.objectName()
-            return True
-        except RuntimeError:
-            return False
-
-    def set_screenshots(self, urls):
-        """Replace the list of cycle-able image URLs. The first entry is also
-        the main thumbnail (used immediately if already cached)."""
-        if not self._is_alive():
-            return
-        urls = [u for u in (urls or []) if u]
-        self._screenshots = urls
-        if urls and urls[0] in self._shot_cache:
-            self._apply_pixmap(self._shot_cache[urls[0]])
-        self._maybe_start_anim()
-        # Proactively validate every URL so broken ones (HTTP 404, network
-        # error, etc.) are dropped from the cycling list up front rather
-        # than producing a black frame when the timer lands on them.
-        self._prefetch_all()
-        # An entry made of nothing but synthetic placeholder URLs has no
-        # real picture: notify the view so it can move the cell to the end
-        # of the grid (the user wants image-bearing items first).
-        if not urls:
-            self._notify_no_image()
-        elif all(isinstance(u, str) and u.startswith("placeholder://")
-                 for u in urls):
-            self._notify_no_image()
-
-    def _prefetch_all(self):
-        if not self._extra_fetch_cb or not self._is_alive():
-            return
-        for url in list(self._screenshots):
-            if url in self._shot_cache:
-                continue
-            def _on_px(pm, _u=url):
-                if not self._is_alive():
-                    return
-                if pm is None or pm.isNull():
-                    self._drop_bad_url(_u)
-                    return
-                self._shot_cache[_u] = pm
-                if self._shot_index < len(self._screenshots) and self._screenshots[self._shot_index] == _u:
-                    self._apply_pixmap(pm)
-            try:
-                self._extra_fetch_cb(url, _on_px)
-            except Exception:
-                self._drop_bad_url(url)
-
-    def set_main_pixmap(self, pm: QPixmap, url: str = ""):
-        if pm is None or pm.isNull():
-            return
-        if not self._is_alive():
-            return
-        if url:
-            self._shot_cache[url] = pm
-            if not self._screenshots:
-                self._screenshots = [url]
-        if self._shot_index < len(self._screenshots):
-            cur_url = self._screenshots[self._shot_index]
-            if not url or cur_url == url:
-                self._apply_pixmap(pm)
-
-    # ---- events --------------------------------------------------------
-
-    def enterEvent(self, ev):
-        self._hovered = True
-        self._info_lbl.setVisible(True)
-        if self._anim_mode_getter() == "hover":
-            # advance to next frame eagerly on hover, then keep cycling
-            if len(self._screenshots) > 1:
-                self._timer.start()
-        super().enterEvent(ev)
-
-    def leaveEvent(self, ev):
-        self._hovered = False
-        self._info_lbl.setVisible(False)
-        if self._anim_mode_getter() == "hover":
-            self._timer.stop()
-            if self._screenshots:
-                self._shot_index = 0
-                self._show_index(0)
-        super().leaveEvent(ev)
-
-    def mousePressEvent(self, ev):
-        if ev.button() == Qt.LeftButton:
-            self.clicked.emit(self._entry)
-        super().mousePressEvent(ev)
-
-    def mouseDoubleClickEvent(self, ev):
-        if ev.button() == Qt.LeftButton:
-            self.dbl_clicked.emit(self._entry)
-        super().mouseDoubleClickEvent(ev)
-
-    def contextMenuEvent(self, ev):
-        if self._context_menu_cb and self._is_alive():
-            self._context_menu_cb(self._entry, ev.globalPos())
-            ev.accept()
-        else:
-            super().contextMenuEvent(ev)
-
-    def resizeEvent(self, ev):
-        # Re-scale the currently shown pixmap to the new width.
-        cur = self._current_pixmap()
-        if cur is not None:
-            self._apply_pixmap(cur)
-        self._position_tag_overlay()
-        self._position_source_overlay()
-        self._position_heart()
-        super().resizeEvent(ev)
-
-    # ---- internals -----------------------------------------------------
-
-    def _apply_style(self):
-        border = "#4fc3f7" if self._selected else "#555"
-        bg = "#1a2733" if self._selected else "#181818"
-        self.setStyleSheet(
-            f"GalleryCell {{ background-color: {bg}; border: 1px solid {border}; border-radius: 4px; }}"
-        )
-
-    def _kickoff_initial_fetch(self):
-        if not self._thumb_fetch_cb:
-            return
-        if not self._is_alive():
-            return
-        self._fetch_attempts += 1
-        try:
-            # Newer callbacks accept an extra `set_tags` callable so that
-            # tags can be derived asynchronously (e.g. from a release lookup).
-            try:
-                self._thumb_fetch_cb(self._entry, self.set_main_pixmap,
-                                     self.set_screenshots, self.set_tags,
-                                     self.set_info_text)
-            except TypeError:
-                try:
-                    self._thumb_fetch_cb(self._entry, self.set_main_pixmap,
-                                         self.set_screenshots, self.set_tags)
-                except TypeError:
-                    self._thumb_fetch_cb(self._entry, self.set_main_pixmap,
-                                         self.set_screenshots)
-        except Exception:
-            pass
-        # Schedule a retry in case this attempt fails silently (the per-gallery
-        # fetch callbacks swallow background errors), leaving the cell stuck on
-        # the "…" placeholder.  The retry is a no-op once a real picture or a
-        # typed placeholder has been shown.
-        self._schedule_retry()
-
-    def _schedule_retry(self):
-        if not self._is_alive():
-            return
-        if self._loaded_ok or self._placeholder_shown:
-            return
-        if self._fetch_attempts >= self._max_fetch_attempts:
-            return
-        # Exponential-ish back-off: 2s, 4s, 8s …
-        delay_ms = 2000 * (2 ** (self._fetch_attempts - 1))
-        try:
-            self._retry_timer.start(delay_ms)
-        except Exception:
-            pass
-
-    def _retry_if_unloaded(self):
-        if not self._is_alive():
-            return
-        if self._loaded_ok or self._placeholder_shown:
-            return
-        self._kickoff_initial_fetch()
-
-    def _maybe_start_anim(self):
-        if len(self._screenshots) <= 1:
-            self._timer.stop()
-            return
-        mode = self._anim_mode_getter()
-        if mode == "timer":
-            self._timer.start()
-        elif mode == "hover" and self._hovered:
-            self._timer.start()
-
-    def _advance(self):
-        if not self._screenshots:
-            return
-        nxt = (self._shot_index + 1) % len(self._screenshots)
-        self._shot_index = nxt
-        self._show_index(nxt)
-
-    def _show_index(self, idx: int):
-        if not self._screenshots:
-            return
-        url = self._screenshots[idx]
-        cached = self._shot_cache.get(url)
-        if cached is not None:
-            self._apply_pixmap(cached)
-            return
-        if self._extra_fetch_cb:
-            def _on_px(pm, _u=url):
-                if pm is None or pm.isNull():
-                    # Drop the broken URL so it never causes a black frame
-                    # in the cycling thumbnail.
-                    self._drop_bad_url(_u)
-                    return
-                self._shot_cache[_u] = pm
-                if self._shot_index < len(self._screenshots) and self._screenshots[self._shot_index] == _u:
-                    self._apply_pixmap(pm)
-            try:
-                self._extra_fetch_cb(url, _on_px)
-            except Exception:
-                self._drop_bad_url(url)
-
-    def _drop_bad_url(self, url: str):
-        """Remove a URL that failed to load from the cycling list and refresh
-        the displayed frame if necessary."""
-        if not self._is_alive():
-            return
-        try:
-            i = self._screenshots.index(url)
-        except ValueError:
-            return
-        del self._screenshots[i]
-        self._shot_cache.pop(url, None)
-        if not self._screenshots:
-            self._timer.stop()
-            self._shot_index = 0
-            self._notify_no_image()
-            return
-        new_idx = self._shot_index
-        if i < self._shot_index:
-            new_idx = self._shot_index - 1
-        if new_idx >= len(self._screenshots):
-            new_idx = 0
-        self._shot_index = new_idx
-        if len(self._screenshots) <= 1:
-            self._timer.stop()
-        # Show whichever URL is now at this slot (cached or trigger a fetch).
-        self._show_index(new_idx)
-
-    def _notify_no_image(self):
-        if self._no_image_notified:
-            return
-        self._no_image_notified = True
-        cb = self._no_image_cb
-        if cb is None:
-            return
-        try:
-            cb(self)
-        except Exception:
-            pass
-
-    def _current_pixmap(self):
-        if not self._screenshots:
-            return None
-        if self._shot_index >= len(self._screenshots):
-            return None
-        return self._shot_cache.get(self._screenshots[self._shot_index])
-
-    def _apply_pixmap(self, pm: QPixmap):
-        if pm is None or pm.isNull():
-            return
-        if not self._is_alive():
-            return
-        # The pixmap width is driven exclusively by the value last requested
-        # by the parent GalleryView (set_thumb_width).  We deliberately ignore
-        # QLabel.width() here: live geometry can momentarily report a much
-        # larger value during a style polish (e.g. after a click toggles the
-        # selection stylesheet) or after returning from a fullscreen viewer,
-        # which would otherwise scale every cell's pixmap to a giant size and
-        # visibly break the 4-column gallery row.
-        target_w = max(40, int(self._thumb_w))
-        # Reserve a 4:3 area, but let the pixmap aspect ratio decide
-        scaled = pm.scaled(target_w, int(target_w * 3 / 4),
-                           Qt.KeepAspectRatio, Qt.SmoothTransformation)
-        self._thumb_lbl.setPixmap(scaled)
-        self._thumb_lbl.setText("")
-        self._position_tag_overlay()
-        self._position_source_overlay()
-        self._position_heart()
-        # A pixmap was successfully applied: mark the cell loaded so the
-        # initial-fetch retry timer stops re-trying.  A synthetic placeholder
-        # counts as "resolved" too (we have nothing better to show), but it is
-        # tracked separately so a real picture is still preferred.
-        cur_url = ""
-        if 0 <= self._shot_index < len(self._screenshots):
-            cur_url = self._screenshots[self._shot_index]
-        is_placeholder = isinstance(cur_url, str) and cur_url.startswith("placeholder://")
-        if is_placeholder:
-            self._placeholder_shown = True
-        else:
-            self._loaded_ok = True
-        try:
-            self._retry_timer.stop()
-        except Exception:
-            pass
-        # Surface the resolved real picture to the parent view so it can be
-        # cached for an instant redraw across an image-first re-sort.
-        cb = self._pixmap_ready_cb
-        if cb is not None:
-            if not is_placeholder:
-                try:
-                    cb(self, pm, cur_url)
-                except Exception:
-                    pass
-
-
-class GalleryView(QWidget):
-    """4-column thumbnail grid wrapper around a QTableWidget.
-
-    The view does NOT own pagination — the surrounding pane drives populate()
-    per page. The grid resizes columns equally; cell widgets are GalleryCell.
-    """
-
-    cell_clicked     = Signal(object)
-    cell_dbl_clicked = Signal(object)
-
-    def __init__(self, rows_per_page_getter, anim_mode_getter,
-                 thumb_fetch_cb, extra_fetch_cb,
-                 title_getter, info_getter, context_menu_cb=None,
-                 tags_getter=None, image_predicate=None,
-                 no_image_key_getter=None,
-                 is_favorite_cb=None, toggle_favorite_cb=None,
-                 source_label_getter=None, tooltip_getter=None,
-                 cols_getter=None, img_size_getter=None, parent=None,
-                 source_overlay_anchor="topleft"):
-        super().__init__(parent)
-        self._rows_per_page_getter = rows_per_page_getter
-        self._anim_mode_getter     = anim_mode_getter
-        self._cols_getter          = cols_getter
-        self._img_size_getter      = img_size_getter
-        self._thumb_fetch_cb       = thumb_fetch_cb
-        self._extra_fetch_cb       = extra_fetch_cb
-        self._title_getter         = title_getter
-        self._info_getter          = info_getter
-        self._tooltip_getter       = tooltip_getter
-        self._context_menu_cb      = context_menu_cb
-        self._tags_getter          = tags_getter or _gallery_extract_tags
-        self._is_favorite_cb       = is_favorite_cb
-        self._toggle_favorite_cb   = toggle_favorite_cb
-        self._source_label_getter  = source_label_getter
-        self._source_overlay_anchor = source_overlay_anchor or "topleft"
-        # Optional predicate(entry) -> bool returning True when the entry
-        # is known to have at least one image at populate-time.  Entries
-        # for which the predicate returns False are moved to the end of
-        # the grid so the user sees items with images first.
-        self._image_predicate      = image_predicate
-        # Optional callable(entry) -> hashable identity used to remember,
-        # across re-populates within the same page, which entries turned
-        # out to have *no* real picture (only a synthetic placeholder).
-        # As background fetches resolve, imageless entries are learned and
-        # the grid is re-sorted (image-bearing first) on a short debounce.
-        self._no_image_key_getter  = no_image_key_getter
-        self._no_image_keys        = set()
-        self._current_entries      = []
-        # Shared pixmap cache (entry-key -> (QPixmap, url)) so an image-first
-        # re-sort can redraw resolved pictures instantly without re-fetching.
-        self._pixmap_cache         = {}
-        self._resort_timer = QTimer(self)
-        self._resort_timer.setSingleShot(True)
-        self._resort_timer.setInterval(250)
-        self._resort_timer.timeout.connect(self._resort_for_images)
-        self._cells = []
-        self._selected_cell = None
-
-        # Let the transparent window background (BackgroundWidget) show through
-        # any empty area of the grid.  The populated cells keep their own
-        # opaque GalleryCell background, so loaded items are NOT transparent;
-        # only the gaps / trailing empty cells reveal the background image.
-        self.setAttribute(Qt.WA_TranslucentBackground, True)
-        self.setStyleSheet("GalleryView { background: transparent; }")
-
-        lay = QVBoxLayout(self)
-        lay.setContentsMargins(0, 0, 0, 0)
-        lay.setSpacing(0)
-
-        self._table = QTableWidget(0, self._cols(), self)
-        self._table.horizontalHeader().setVisible(False)
-        self._table.verticalHeader().setVisible(False)
-        self._table.setSelectionMode(QAbstractItemView.NoSelection)
-        self._table.setEditTriggers(QAbstractItemView.NoEditTriggers)
-        self._table.setFocusPolicy(Qt.NoFocus)
-        self._table.setShowGrid(False)
-        self._table.horizontalHeader().setSectionResizeMode(QHeaderView.Stretch)
-        self._table.horizontalHeader().setStretchLastSection(True)
-        self._table.verticalHeader().setSectionResizeMode(QHeaderView.Fixed)
-        # Make the table and its viewport transparent so empty cells reveal
-        # the background image instead of an opaque table background.
-        self._table.setAttribute(Qt.WA_TranslucentBackground, True)
-        self._table.viewport().setAttribute(Qt.WA_TranslucentBackground, True)
-        self._table.viewport().setAutoFillBackground(False)
-        self._table.setFrameShape(QFrame.NoFrame)
-        self._table.setStyleSheet(
-            "QTableWidget { background: transparent; border: none; }"
-            "QTableWidget::viewport { background: transparent; }"
-        )
-        lay.addWidget(self._table)
-
-        self._table.viewport().installEventFilter(self)
-
-    def page_size(self) -> int:
-        try:
-            n = int(self._rows_per_page_getter())
-        except Exception:
-            n = DEFAULT_GALLERY_ROWS_PER_PAGE
-        n = max(GALLERY_MIN_ROWS, min(GALLERY_MAX_ROWS, n))
-        return n * self._cols()
-
-    def _cols(self) -> int:
-        """Return the current number of gallery columns."""
-        if self._cols_getter is not None:
-            try:
-                return int(self._cols_getter())
-            except Exception:
-                pass
-        return DEFAULT_GALLERY_COLS
-
-    def _row_h(self, col_w: int) -> int:
-        """Return row height for *col_w* scaled by the current image-size setting."""
-        size = ""
-        if self._img_size_getter is not None:
-            try:
-                size = str(self._img_size_getter()).lower()
-            except Exception:
-                pass
-        if size == "small":
-            return int(col_w * 3 / 8) + 40
-        if size == "large":
-            return int(col_w * 3 / 2) + 72
-        return int(col_w * 3 / 4) + 56  # medium (default)
-
-    def _entry_image_key(self, e):
-        """Return a hashable identity for *e* used to track imageless entries
-        and cache resolved pixmaps.
-
-        When no explicit key getter is supplied we fall back to the Python
-        object identity of the entry dict.  That is stable for the lifetime of
-        a populated page (the same dict instances are reused when the grid is
-        re-sorted) and the per-page caches are reset on every fresh
-        ``populate``, so stale identities can never leak across pages."""
-        if self._no_image_key_getter is not None:
-            try:
-                k = self._no_image_key_getter(e)
-                if k is not None:
-                    return k
-            except Exception:
-                pass
-        try:
-            return id(e)
-        except Exception:
-            return None
-
-    def _has_image(self, e) -> bool:
-        """Best-effort 'does this entry have a real picture' decision.
-
-        Combines the populate-time predicate (when supplied) with what we
-        have *learned* at runtime: an entry whose key is in the no-image set
-        is treated as imageless regardless of the predicate."""
-        key = self._entry_image_key(e)
-        if key is not None and key in self._no_image_keys:
-            return False
-        if self._image_predicate is not None:
-            try:
-                return bool(self._image_predicate(e))
-            except Exception:
-                return True
-        return True
-
-    def _order_image_first(self, entries):
-        """Stable-partition *entries* so items with a real picture come first
-        and known-imageless items (placeholder bitmap/file/text/utility) sink
-        to the bottom."""
-        if not entries:
-            return entries
-        if self._image_predicate is None and not self._no_image_keys:
-            return entries
-        with_img    = [e for e in entries if self._has_image(e)]
-        without_img = [e for e in entries if not self._has_image(e)]
-        return with_img + without_img
-
-    def populate(self, entries):
-        """Render `entries` (list of dicts) into the grid. Excess capacity is
-        cleared. Caller must have already paged the entries appropriately."""
-        entries = list(entries or [])
-        # A fresh page of entries: forget what we learned about the previous
-        # page so the new content is evaluated from scratch.
-        self._no_image_keys = set()
-        self._pixmap_cache = {}
-        self._resort_timer.stop()
-        entries = self._order_image_first(entries)
-        self._current_entries = list(entries)
-        self._render_entries(entries)
-
-    def _render_entries(self, entries):
-        """(Re)build the grid widgets for *entries* in the given order.  Used
-        by ``populate`` for a fresh page and by ``_resort_for_images`` when the
-        image-first ordering changes at runtime."""
-        entries = list(entries or [])
-        rows_needed = (len(entries) + self._cols() - 1) // self._cols() if entries else 0
-        # Tear down any existing cells
-        for c in self._cells:
-            try:
-                c.setParent(None)
-                c.deleteLater()
-            except Exception:
-                pass
-        self._cells = []
-        self._selected_cell = None
-        self._table.setColumnCount(self._cols())
-        self._table.clearContents()
-        self._table.setRowCount(rows_needed)
-
-        # Compute a sane row height: thumbnail size driven by image-size setting.
-        # We re-apply this on resize as well.
-        vp_w = max(200, self._table.viewport().width())
-        col_w = max(80, vp_w // self._cols() - 6)
-        row_h = self._row_h(col_w)
-        for r in range(rows_needed):
-            self._table.setRowHeight(r, row_h)
-        # Ensure the table is always tall enough to show 4 rows without scrolling.
-        self._table.setMinimumHeight(row_h * 4)
-
-        for i, e in enumerate(entries):
-            r, c = divmod(i, self._cols())
-            title   = self._title_getter(e) if self._title_getter   else ""
-            info    = self._info_getter(e)  if self._info_getter    else ""
-            tooltip = self._tooltip_getter(e) if self._tooltip_getter else ""
-            tags  = []
-            if self._tags_getter:
-                try:
-                    tags = list(self._tags_getter(e) or [])
-                except Exception:
-                    tags = []
-            cell = GalleryCell(
-                entry=e,
-                anim_mode_getter=self._anim_mode_getter,
-                thumb_fetch_cb=self._thumb_fetch_cb,
-                extra_fetch_cb=self._extra_fetch_cb,
-                title_text=title,
-                info_text=info,
-                tooltip_text=tooltip,
-                context_menu_cb=self._context_menu_cb,
-                tags=tags,
-                parent=self._table,
-                is_favorite_cb=self._is_favorite_cb,
-                toggle_favorite_cb=self._toggle_favorite_cb,
-                source_label_getter=self._source_label_getter,
-                source_overlay_anchor=self._source_overlay_anchor,
-            )
-            cell.set_thumb_width(col_w)
-            cell.clicked.connect(self._on_cell_clicked)
-            cell.dbl_clicked.connect(self._on_cell_dbl_clicked)
-            cell._no_image_cb = self._on_cell_no_image
-            cell._pixmap_ready_cb = self._on_cell_pixmap_ready
-            # Seed from the shared pixmap cache so a cell that already resolved
-            # its picture before a re-sort redraws instantly (no re-download,
-            # no flicker) instead of falling back to the "…" placeholder.
-            key = self._entry_image_key(e)
-            if key is not None:
-                cached = self._pixmap_cache.get(key)
-                if cached is not None:
-                    cell.set_main_pixmap(cached[0], cached[1])
-            self._table.setCellWidget(r, c, cell)
-            self._cells.append(cell)
-
-    def _on_cell_pixmap_ready(self, cell, pm, url):
-        """Cache a cell's resolved picture (keyed by entry identity) so a later
-        image-first re-sort can redraw it instantly without re-fetching."""
-        try:
-            if pm is None or pm.isNull():
-                return
-            entry = cell.entry()
-        except Exception:
-            return
-        key = self._entry_image_key(entry)
-        if key is None:
-            return
-        if isinstance(url, str) and url.startswith("placeholder://"):
-            return
-        self._pixmap_cache[key] = (pm, url or "")
-
-    def select_entry(self, predicate):
-        """Mark the first cell whose entry satisfies `predicate(entry)` as
-        selected. Pass `lambda e: False` to clear."""
-        for cell in self._cells:
-            sel = False
-            try:
-                sel = bool(predicate(cell.entry()))
-            except Exception:
-                sel = False
-            cell.set_selected(sel)
-            if sel:
-                self._selected_cell = cell
-
-    def eventFilter(self, obj, ev):
-        if ev.type() == QEvent.Resize and obj is self._table.viewport():
-            self._apply_dimensions()
-        return super().eventFilter(obj, ev)
-
-    def refresh_favorites(self):
-        """Re-query the favorite predicate on every cell. Called when the
-        favorite state changes externally (e.g. via the fullscreen viewer
-        or another pane's favorite toggle)."""
-        for cell in self._cells:
-            try:
-                cell.set_favorite_state_changed()
-            except Exception:
-                pass
-
-    def showEvent(self, ev):
-        # When the gallery becomes visible again — for example after the user
-        # closes the in-pane fullscreen viewer that lives in the same
-        # QStackedWidget — re-apply dimensions so every cell rescales its
-        # thumbnail to the current column width.  Without this, a cell's
-        # QPixmap can remain at the size it was rendered at right before the
-        # viewer was shown and visibly overflow the 4-column row.
-        super().showEvent(ev)
-        QTimer.singleShot(0, self._apply_dimensions)
-
-    def _apply_dimensions(self):
-        cols  = self._cols()
-        if self._table.columnCount() != cols:
-            self._table.setColumnCount(cols)
-        vp_w = max(200, self._table.viewport().width())
-        col_w = max(80, vp_w // cols - 6)
-        row_h = self._row_h(col_w)
-        for r in range(self._table.rowCount()):
-            self._table.setRowHeight(r, row_h)
-        for cell in self._cells:
-            cell.set_thumb_width(col_w)
-        # Keep the table tall enough to show 4 full rows without scrolling.
-        self._table.setMinimumHeight(row_h * 4)
-
-    def _on_cell_clicked(self, entry):
-        sender = self.sender()
-        for cell in self._cells:
-            cell.set_selected(cell is sender)
-        self._selected_cell = sender
-        self.cell_clicked.emit(entry)
-
-    def _on_cell_dbl_clicked(self, entry):
-        self.cell_dbl_clicked.emit(entry)
-
-    def _on_cell_no_image(self, cell):
-        """Called by a GalleryCell when it has determined it has no usable
-        image (an explicit empty screenshot list, only placeholder URLs, or
-        every real URL failed to load).
-
-        We remember the entry's identity in ``self._no_image_keys`` and arm a
-        short debounce timer.  When it fires, the whole page is re-sorted
-        (image-bearing first) via a full re-populate — the only safe way to
-        reorder ``QTableWidget`` cell widgets without Qt deleting them.  The
-        debounce coalesces the burst of callbacks that arrive as a page's
-        fetches resolve into a single relayout, and the shared pixmap cache
-        means image-bearing cells redraw instantly with no re-download."""
-        try:
-            entry = cell.entry()
-        except Exception:
-            entry = None
-        if entry is None:
-            return
-        key = self._entry_image_key(entry)
-        if key is None:
-            return
-        if key in self._no_image_keys:
-            return
-        self._no_image_keys.add(key)
-        # Only bother re-sorting if this imageless cell is not already trailing
-        # the grid (i.e. there is at least one later cell that *does* have an
-        # image and should move ahead of it).
-        self._resort_timer.start()
-
-    def _resort_for_images(self):
-        """Re-sort the current page so image-bearing entries lead, then
-        re-populate.  No-op when the order would not change."""
-        if not self._current_entries:
-            return
-        reordered = self._order_image_first(self._current_entries)
-        if [id(e) for e in reordered] == [id(e) for e in self._current_entries]:
-            return
-        # Preserve the current selection across the relayout.
-        sel_entry = None
-        if self._selected_cell is not None:
-            try:
-                sel_entry = self._selected_cell.entry()
-            except Exception:
-                sel_entry = None
-        self._render_entries(reordered)
-        self._current_entries = list(reordered)
-        if sel_entry is not None:
-            self.select_entry(lambda e: e is sel_entry)
-
-    def _apply_no_image_relayout(self):
-        # Retained for backwards compatibility; routes to the debounced
-        # image-first re-sort.
-        self._resort_timer.start()
-
-    def _relayout_cells(self):
-        """Re-bind every cell in self._cells to its new (row, col) slot."""
-        cols = self._cols()
-        rows_needed = (len(self._cells) + cols - 1) // cols if self._cells else 0
-        # Detach existing cells from their current slots without deleting
-        # them.  setCellWidget(r,c,None) would delete the previously set
-        # widget, so we use removeCellWidget which only detaches.
-        total_rows = self._table.rowCount()
-        for r in range(total_rows):
-            for c in range(self._table.columnCount()):
-                if self._table.cellWidget(r, c) is not None:
-                    self._table.removeCellWidget(r, c)
-        self._table.setColumnCount(cols)
-        self._table.setRowCount(rows_needed)
-        # Re-apply current row height to all rows.
-        vp_w = max(200, self._table.viewport().width())
-        col_w = max(80, vp_w // cols - 6)
-        row_h = self._row_h(col_w)
-        for r in range(rows_needed):
-            self._table.setRowHeight(r, row_h)
-        for i, cell in enumerate(self._cells):
-            r, c = divmod(i, cols)
-            self._table.setCellWidget(r, c, cell)
-
-
-def _gallery_stars(rating_value) -> str:
-    """Return a 5-star unicode string for a 0–10 or 0–5 rating value."""
-    try:
-        v = float(rating_value)
-    except (TypeError, ValueError):
-        return ""
-    # Normalise: values > 5 are assumed to be on a 0–10 scale
-    if v > 5:
-        v = v / 2.0
-    v = max(0.0, min(5.0, v))
-    full  = int(v + 0.5)
-    empty = 5 - full
-    return "★" * full + "☆" * empty + f"  ({rating_value})"
-
-
-class GalleryItemViewer(QWidget):
-    """In-pane item viewer opened from Gallery mode (double-click on a cell).
-
-    Displayed inside the pane's QStackedWidget (index 2) rather than as a
-    separate OS window, so it fills the same area as the normal pane view.
-
-    Layout
-    ------
-    Left (stretch 3)  – large screenshot, auto-cycling slideshow, ◀/▶ nav
-    Right (fixed 340) – scrollable metadata (title, key/value rows, star
-                        rating) + action bar (Download / Send to SD / NextSync)
-
-    Close: ✕ button (top-right) or Escape key → pops stack back to index 0.
-    """
-
-    _BTN_STYLE = (
-        "QPushButton { color: #eee; background: #2a2a2a; border: 1px solid #444;"
-        " border-radius: 4px; padding: 6px 12px; text-align: left; }"
-        "QPushButton:hover { background: #3a3a3a; border-color: #666; }"
-        "QPushButton:disabled { color: #555; background: #1a1a1a; border-color: #333; }"
-    )
-
-    def __init__(self, title: str, info_rows: list, screenshots: list,
-                 extra_fetch_cb, tags=None, parent=None):
-        """
-        Parameters
-        ----------
-        title          : program name shown at top of metadata panel
-        info_rows      : list of (label, value) tuples; empty value → skip row
-        screenshots    : list of image URL strings for the slideshow
-        extra_fetch_cb : callable(url, on_pixmap_cb) – async image loader
-        """
-        super().__init__(parent)
-        self.setFocusPolicy(Qt.StrongFocus)
-        self.setStyleSheet("background: #0d0d0d;")
-
-        self._screenshots    = list(screenshots or [])
-        self._shot_index     = 0
-        self._shot_cache     = {}
-        self._extra_fetch_cb = extra_fetch_cb
-        self._close_fn       = None   # set by install_into_stack()
-        self._tags           = [str(t) for t in (tags or []) if t]
-        self._is_favorite_cb     = None
-        self._toggle_favorite_cb = None
-        self._fav_entry          = None
-        self._title              = title or ""
-        self._placeholder_label    = ""
-        self._placeholder_subtitle = ""
-
-        # ── root layout ──────────────────────────────────────────────────
-        root = QHBoxLayout(self)
-        root.setContentsMargins(0, 0, 0, 0)
-        root.setSpacing(0)
-
-        # ── LEFT: image area ─────────────────────────────────────────────
-        img_panel = QWidget()
-        img_panel.setStyleSheet("background: #0a0a0a;")
-        img_layout = QVBoxLayout(img_panel)
-        img_layout.setContentsMargins(8, 8, 8, 8)
-        img_layout.setSpacing(4)
-
-        self._img_lbl = QLabel()
-        self._img_lbl.setAlignment(Qt.AlignCenter)
-        self._img_lbl.setStyleSheet("background: #0a0a0a; color: #666;")
-        self._img_lbl.setText("Loading…")
-        self._img_lbl.setSizePolicy(QSizePolicy.Expanding, QSizePolicy.Expanding)
-        self._img_lbl.setCursor(Qt.PointingHandCursor)
-        img_layout.addWidget(self._img_lbl, 1)
-
-        # Tag overlay floating in the top-right corner of the image area.
-        self._tag_overlay = QLabel(self._img_lbl)
-        self._tag_overlay.setAttribute(Qt.WA_TransparentForMouseEvents, True)
-        self._tag_overlay.setAlignment(Qt.AlignTop | Qt.AlignRight)
-        self._tag_overlay.setTextFormat(Qt.RichText)
-        self._tag_overlay.setStyleSheet("background: transparent;")
-        self._tag_overlay.setVisible(False)
-        self._img_lbl.installEventFilter(self)
-        self._refresh_tag_overlay()
-
-        # ◀ counter ▶
-        nav_row = QHBoxLayout()
-        _nav_ss = (
-            "QToolButton { color: white; background: #2b2b2b; border: none;"
-            " font-size: 20px; padding: 5px 14px; }"
-            "QToolButton:hover { background: #484848; }"
-        )
-        self._prev_btn = QToolButton()
-        self._prev_btn.setText("◀")
-        self._prev_btn.setStyleSheet(_nav_ss)
-        self._next_btn = QToolButton()
-        self._next_btn.setText("▶")
-        self._next_btn.setStyleSheet(_nav_ss)
-        self._shot_counter = QLabel("")
-        self._shot_counter.setAlignment(Qt.AlignCenter)
-        self._shot_counter.setStyleSheet("color: #777; font-size: 11px; min-width: 60px;")
-        nav_row.addStretch()
-        nav_row.addWidget(self._prev_btn)
-        nav_row.addWidget(self._shot_counter)
-        nav_row.addWidget(self._next_btn)
-        nav_row.addStretch()
-        img_layout.addLayout(nav_row)
-
-        self._prev_btn.clicked.connect(self._go_prev)
-        self._next_btn.clicked.connect(self._go_next)
-
-        root.addWidget(img_panel, 3)
-
-        # ── RIGHT: metadata + actions ────────────────────────────────────
-        right_panel = QWidget()
-        right_panel.setStyleSheet("background: #111;")
-        right_panel.setMinimumWidth(300)
-        right_panel.setMaximumWidth(400)
-        right_layout = QVBoxLayout(right_panel)
-        right_layout.setContentsMargins(0, 0, 0, 0)
-        right_layout.setSpacing(0)
-
-        # ── close bar ────────────────────────────────────────────────────
-        close_bar = QHBoxLayout()
-        close_bar.setContentsMargins(10, 8, 10, 4)
-        close_bar.addStretch()
-        self._heart_btn = QToolButton()
-        self._heart_btn.setText("♡")
-        self._heart_btn.setCursor(Qt.PointingHandCursor)
-        self._heart_btn.setStyleSheet(
-            "QToolButton { color: #ff5577; background: #2b2b2b; border: none;"
-            " font-size: 17px; padding: 3px 10px; }"
-            "QToolButton:hover { background: #4a2a35; }"
-        )
-        self._heart_btn.setVisible(False)
-        self._heart_btn.clicked.connect(self._on_heart_clicked)
-        close_bar.addWidget(self._heart_btn)
-        self._close_btn = QToolButton()
-        self._close_btn.setText("✕")
-        self._close_btn.setStyleSheet(
-            "QToolButton { color: white; background: #2b2b2b; border: none;"
-            " font-size: 15px; padding: 3px 10px; }"
-            "QToolButton:hover { background: #c00; }"
-        )
-        self._close_btn.clicked.connect(self._do_close)
-        close_bar.addWidget(self._close_btn)
-        right_layout.addLayout(close_bar)
-
-        # ── scrollable metadata ───────────────────────────────────────────
-        self._scroll = QScrollArea()
-        self._scroll.setWidgetResizable(True)
-        # Long Description values would otherwise let the inner widget grow
-        # horizontally and shift the row labels (e.g. "Description:") under
-        # the left image area.  Forbid the horizontal scrollbar so the
-        # content has to wrap inside the available right-panel width.
-        self._scroll.setHorizontalScrollBarPolicy(Qt.ScrollBarAlwaysOff)
-        self._scroll.setStyleSheet("QScrollArea { border: none; background: #111; }")
-        self._meta_widget = QWidget()
-        self._rebuild_meta(title, info_rows)
-        right_layout.addWidget(self._scroll, 1)
-
-        # ── action bar ────────────────────────────────────────────────────
-        action_bar = QWidget()
-        action_bar.setStyleSheet("background: #0d0d0d; border-top: 1px solid #2a2a2a;")
-        ab_layout = QVBoxLayout(action_bar)
-        ab_layout.setContentsMargins(10, 8, 10, 10)
-        ab_layout.setSpacing(6)
-
-        self.btn_open_web = QPushButton("🌐  Open on website")
-        self.btn_download = QPushButton("⬇  Download")
-        self.btn_send_sd  = QPushButton("💾  Send to SD card")
-        self.btn_send_ns  = QPushButton("🔁  Send via NextSync")
-        self.btn_open_web.setStyleSheet(self._BTN_STYLE)
-        self.btn_open_web.setEnabled(False)
-        self.btn_open_web.setVisible(False)
-        ab_layout.addWidget(self.btn_open_web)
-        for btn in (self.btn_download, self.btn_send_sd, self.btn_send_ns):
-            btn.setStyleSheet(self._BTN_STYLE)
-            btn.setEnabled(False)
-            ab_layout.addWidget(btn)
-
-        right_layout.addWidget(action_bar)
-        root.addWidget(right_panel, 0)
-
-        # ── slideshow timer ───────────────────────────────────────────────
-        self._timer = QTimer(self)
-        self._timer.setInterval(4000)
-        self._timer.timeout.connect(self._go_next)
-
-        self._update_nav()
-        if self._screenshots:
-            self._show_index(0)
-            # Proactively validate every URL so broken ones are pruned
-            # before the cycling timer reaches them.
-            self._prefetch_all()
-
-    # ── public API ────────────────────────────────────────────────────────
-
-    def install_into_stack(self, stack: QStackedWidget, close_fn=None):
-        """Add this widget to *stack* (if not already there) and show it."""
-        self._close_fn = close_fn
-        if stack.indexOf(self) == -1:
-            stack.addWidget(self)
-        stack.setCurrentWidget(self)
-        self.setFocus()
-        if len(self._screenshots) > 1:
-            self._timer.start()
-
-    def set_screenshots(self, urls: list):
-        """Replace screenshot list and restart slideshow."""
-        self._timer.stop()
-        self._screenshots = list(urls or [])
-        self._shot_index  = 0
-        self._shot_cache  = {}
-        self._update_nav()
-        if self._screenshots:
-            self._show_index(0)
-            if len(self._screenshots) > 1:
-                self._timer.start()
-            # Proactively fetch every URL so that broken ones (HTTP 404,
-            # network error, etc.) are pruned up front rather than only
-            # when the cycling timer happens to land on them.
-            self._prefetch_all()
-        else:
-            self._render_placeholder()
-
-    def _prefetch_all(self):
-        """Fire an async fetch for every URL in the cycling list so that
-        unreachable images get removed from the list ASAP."""
-        if not self._extra_fetch_cb:
-            return
-        for url in list(self._screenshots):
-            if url in self._shot_cache:
-                continue
-            def _on_px(pm, _u=url):
-                if pm is None or pm.isNull():
-                    self._drop_bad_url(_u)
-                    return
-                self._shot_cache[_u] = pm
-                # If this URL is currently displayed, refresh the view.
-                if self._screenshots and self._shot_index < len(self._screenshots) \
-                        and self._screenshots[self._shot_index] == _u:
-                    self._display_pixmap(pm)
-            try:
-                self._extra_fetch_cb(url, _on_px)
-            except Exception:
-                self._drop_bad_url(url)
-
-    def refresh_meta(self, title: str, rows: list):
-        """Rebuild the metadata scroll panel (called from async callbacks)."""
-        self._rebuild_meta(title, rows)
-
-    def set_tags(self, tags):
-        """Replace the overlay tag badges shown over the screenshot area."""
-        self._tags = [str(t) for t in (tags or []) if t]
-        self._refresh_tag_overlay()
-
-    def set_favorite_hooks(self, entry, is_favorite_cb, toggle_favorite_cb):
-        """Wire up the heart control to a (de)favorite callback for *entry*."""
-        self._fav_entry          = entry
-        self._is_favorite_cb     = is_favorite_cb
-        self._toggle_favorite_cb = toggle_favorite_cb
-        self._refresh_heart()
-
-    def _refresh_heart(self):
-        btn = getattr(self, "_heart_btn", None)
-        if btn is None:
-            return
-        if not self._toggle_favorite_cb or self._fav_entry is None:
-            btn.setVisible(False)
-            return
-        is_fav = False
-        if self._is_favorite_cb is not None:
-            try:
-                is_fav = bool(self._is_favorite_cb(self._fav_entry))
-            except Exception:
-                is_fav = False
-        btn.setText("♥" if is_fav else "♡")
-        btn.setToolTip("Remove from favorites" if is_fav else "Add to favorites")
-        btn.setVisible(True)
-
-    def _on_heart_clicked(self):
-        if not self._toggle_favorite_cb or self._fav_entry is None:
-            return
-        try:
-            self._toggle_favorite_cb(self._fav_entry)
-        except Exception:
-            pass
-        self._refresh_heart()
-
-    def _refresh_tag_overlay(self):
-        ov = getattr(self, "_tag_overlay", None)
-        if ov is None:
-            return
-        if not self._tags:
-            ov.setVisible(False)
-            ov.setText("")
-            return
-        chip_css = (
-            "background:#1c3a52;color:#bfe6ff;border:1px solid #2f6f9a;"
-            "border-radius:3px;padding:2px 7px;margin:2px 3px;"
-            "font-size:10pt;font-weight:600;"
-        )
-        chips = "&nbsp;".join(
-            f"<span style='{chip_css}'>{t}</span>" for t in self._tags
-        )
-        ov.setText(f"<div style='text-align:right;'>{chips}</div>")
-        ov.setVisible(True)
-        self._position_tag_overlay()
-
-    def _position_tag_overlay(self):
-        ov = getattr(self, "_tag_overlay", None)
-        if ov is None or not ov.isVisible():
-            return
-        pad = 8
-        max_w = max(60, self._img_lbl.width() - 2 * pad)
-        ov.adjustSize()
-        w = min(ov.width(), max_w)
-        h = ov.sizeHint().height()
-        ov.setGeometry(self._img_lbl.width() - w - pad, pad, w, h)
-        ov.raise_()
-
-    def eventFilter(self, obj, ev):
-        if obj is self._img_lbl:
-            if ev.type() == QEvent.Resize:
-                self._position_tag_overlay()
-            elif ev.type() in (QEvent.MouseButtonPress, QEvent.MouseButtonDblClick):
-                if ev.button() == Qt.LeftButton:
-                    self._do_close()
-                    return True
-        return super().eventFilter(obj, ev)
-
-    def set_actions(self, download_cb=None, send_sd_cb=None, send_ns_cb=None,
-                    sd_enabled=False, ns_enabled=False,
-                    sd_tooltip="", ns_tooltip=""):
-        """Wire action buttons.  Pass None to keep a button disabled."""
-        self._wire_btn(self.btn_download, download_cb, True)
-        self._wire_btn(self.btn_send_sd,  send_sd_cb,  sd_enabled, sd_tooltip)
-        self._wire_btn(self.btn_send_ns,  send_ns_cb,  ns_enabled, ns_tooltip)
-
-    def set_open_web_url(self, url: str, site_label: str = ""):
-        """Wire (and show) the 'Open on website' button.  The button opens
-        *url* in the user's default external browser via ``webbrowser.open``.
-        When *url* is empty the button is hidden."""
-        url = (url or "").strip()
-        try:
-            with __import__("warnings").catch_warnings():
-                __import__("warnings").simplefilter("ignore", RuntimeWarning)
-                self.btn_open_web.clicked.disconnect()
-        except (RuntimeError, TypeError):
-            pass
-        if not url:
-            self.btn_open_web.setVisible(False)
-            self.btn_open_web.setEnabled(False)
-            return
-        label = f"🌐  Open on {site_label}" if site_label else "🌐  Open on website"
-        self.btn_open_web.setText(label)
-        self.btn_open_web.setToolTip(url)
-        self.btn_open_web.setVisible(True)
-        self.btn_open_web.setEnabled(True)
-        def _go(_=False, _u=url):
-            try:
-                webbrowser.open(_u, new=2)
-            except Exception:
-                pass
-        self.btn_open_web.clicked.connect(_go)
-
-    def set_download_available(self, has_dl: bool):
-        """Show or hide all three action buttons depending on whether any
-        downloadable files exist for this entry."""
-        for btn in (self.btn_download, self.btn_send_sd, self.btn_send_ns):
-            btn.setVisible(bool(has_dl))
-
-    # ── private helpers ───────────────────────────────────────────────────
-
-    def _wire_btn(self, btn, cb, enabled, tooltip=""):
-        import warnings
-        try:
-            with warnings.catch_warnings():
-                warnings.simplefilter("ignore", RuntimeWarning)
-                btn.clicked.disconnect()
-        except (RuntimeError, TypeError):
-            pass
-        btn.setEnabled(bool(enabled) and cb is not None)
-        if tooltip:
-            btn.setToolTip(tooltip)
-        if cb is not None:
-            btn.clicked.connect(cb)
-
-    def set_placeholder(self, label: str, subtitle: str = ""):
-        """Store the label/subtitle for the placeholder shown when no screenshots
-        are available, and render it immediately if the image area is empty."""
-        self._placeholder_label    = str(label or "")
-        self._placeholder_subtitle = str(subtitle or "") or self._title
-        if not self._screenshots:
-            self._render_placeholder()
-
-    def _render_placeholder(self):
-        """Render a typed placeholder pixmap in the image area (same style as
-        gallery thumbnails: yellow label + subtitle on dark background)."""
-        label    = self._placeholder_label or "FILE"
-        subtitle = self._placeholder_subtitle or self._title
-        pm = zxfmt_make_placeholder_pixmap(label, subtitle)
-        if pm and not pm.isNull():
-            sz = self._img_lbl.size()
-            if sz.width() > 0 and sz.height() > 0:
-                self._img_lbl.setPixmap(
-                    pm.scaled(sz, Qt.KeepAspectRatio, Qt.SmoothTransformation)
-                )
-            else:
-                self._img_lbl.setPixmap(pm)
-            self._img_lbl.setText("")
-        else:
-            self._img_lbl.setPixmap(QPixmap())
-            self._img_lbl.setText("No preview available")
-        self._position_tag_overlay()
-
-    def _rebuild_meta(self, title: str, rows: list):
-        meta_widget = QWidget()
-        meta_widget.setStyleSheet("background: #111;")
-        # Constrain the inner widget so it can never grow wider than the
-        # right panel; combined with setMinimumWidth(0) on each row label
-        # this forces text to wrap rather than extend off-screen.
-        meta_widget.setMaximumWidth(380)
-        ml = QVBoxLayout(meta_widget)
-        ml.setContentsMargins(16, 12, 16, 12)
-        ml.setSpacing(6)
-
-        title_lbl = QLabel(title or "")
-        tf = title_lbl.font()
-        _tps = tf.pointSize()
-        if _tps <= 0:
-            _tps = max(QFontInfo(tf).pointSize(), 9)
-        tf.setPointSize(_tps + 4)
-        tf.setBold(True)
-        title_lbl.setFont(tf)
-        title_lbl.setWordWrap(True)
-        title_lbl.setStyleSheet("color: #fff; margin-bottom: 8px;")
-        ml.addWidget(title_lbl)
-
-        sep = QFrame()
-        sep.setFrameShape(QFrame.HLine)
-        sep.setStyleSheet("color: #2e2e2e;")
-        ml.addWidget(sep)
-
-        import html as _html
-        import re as _re
-
-        for row in (rows or []):
-            label, value = row[0], row[1]
-            is_html = row[2] if len(row) > 2 else False
-            if not value:
-                continue
-
-            # Always sanitise the value: strip every HTML tag and decode
-            # entities before re-escaping. All rows are rendered inside a
-            # Qt.RichText label, so escaped entities (e.g. &lt;LI&gt;) would
-            # be decoded by Qt and show as visible tags. The is_html flag
-            # merely signals that the raw value *may* contain HTML markup;
-            # the plain-text path must apply the same treatment because any
-            # value sourced from an API can carry incidental HTML.
-            raw = str(value or "")
-            raw = _re.sub(r"<br\s*/?>", "\n", raw, flags=_re.IGNORECASE)
-            raw = _re.sub(r"</p\s*>",   "\n\n", raw, flags=_re.IGNORECASE)
-            raw = _re.sub(r"<li\s*/?>", "\n• ", raw, flags=_re.IGNORECASE)
-            raw = _re.sub(r"<[^>]+>",   "",    raw)
-            raw = _html.unescape(raw)
-            # Collapse excessive blank runs.
-            raw = _re.sub(r"\n{3,}", "\n\n", raw).strip()
-            val_html = _html.escape(raw).replace("\n", "<br>")
-
-            row_lbl = QLabel(
-                '<span style="color:#888; font-weight:bold;">'
-                + _html.escape(str(label)) + '</span><br>'
-                '<span style="color:#ddd; word-wrap:break-word; '
-                'word-break:break-all;">' + val_html + '</span>'
-            )
-            row_lbl.setTextFormat(Qt.RichText)
-            row_lbl.setWordWrap(True)
-            row_lbl.setAlignment(Qt.AlignTop | Qt.AlignLeft)
-            row_lbl.setTextInteractionFlags(Qt.TextSelectableByMouse)
-            row_lbl.setOpenExternalLinks(True)
-            # Critical: Ignored horizontal policy + minimum width 0 means
-            # the label cannot inflate to its unwrapped sizeHint() width,
-            # so long unbroken text MUST wrap inside the available space.
-            row_lbl.setSizePolicy(QSizePolicy.Ignored, QSizePolicy.MinimumExpanding)
-            row_lbl.setMinimumWidth(0)
-            row_lbl.setMaximumWidth(340)
-            row_lbl.setContentsMargins(0, 2, 0, 6)
-            ml.addWidget(row_lbl)
-
-        ml.addStretch()
-        old = self._scroll.widget()
-        self._scroll.setWidget(meta_widget)
-        if old:
-            old.deleteLater()
-        self._meta_widget = meta_widget
-
-    def _do_close(self):
-        self._timer.stop()
-        if self._close_fn:
-            self._close_fn()
-
-    # ── keyboard ──────────────────────────────────────────────────────────
-
-    def keyPressEvent(self, ev):
-        if ev.key() == Qt.Key_Escape:
-            self._do_close()
-        elif ev.key() == Qt.Key_Left:
-            self._go_prev()
-        elif ev.key() == Qt.Key_Right:
-            self._go_next()
-        else:
-            super().keyPressEvent(ev)
-
-    # ── navigation ────────────────────────────────────────────────────────
-
-    def _go_prev(self):
-        if not self._screenshots:
-            return
-        self._timer.stop()
-        self._shot_index = (self._shot_index - 1) % len(self._screenshots)
-        self._show_index(self._shot_index)
-        self._timer.start()
-
-    def _go_next(self):
-        if not self._screenshots:
-            return
-        self._timer.stop()
-        self._shot_index = (self._shot_index + 1) % len(self._screenshots)
-        self._show_index(self._shot_index)
-        self._timer.start()
-
-    def _update_nav(self):
-        multi = len(self._screenshots) > 1
-        self._prev_btn.setVisible(multi)
-        self._next_btn.setVisible(multi)
-        self._shot_counter.setText(
-            f"{self._shot_index + 1} / {len(self._screenshots)}" if self._screenshots else ""
-        )
-
-    def _show_index(self, idx: int):
-        if not self._screenshots:
-            return
-        if idx < 0 or idx >= len(self._screenshots):
-            idx = 0
-        url = self._screenshots[idx]
-        self._shot_index = idx
-        self._update_nav()
-        cached = self._shot_cache.get(url)
-        if cached is not None:
-            self._display_pixmap(cached)
-            return
-        self._img_lbl.setPixmap(QPixmap())
-        self._img_lbl.setText("Loading…")
-        if self._extra_fetch_cb:
-            def _on_px(pm, _u=url):
-                if pm is None or pm.isNull():
-                    # The image failed to download (e.g. HTTP 404).
-                    # Drop it from the cycling list so the slideshow
-                    # never shows a black frame for that slot, and
-                    # advance to a usable neighbour.
-                    self._drop_bad_url(_u)
-                    return
-                self._shot_cache[_u] = pm
-                if self._screenshots and self._shot_index < len(self._screenshots) \
-                        and self._screenshots[self._shot_index] == _u:
-                    self._display_pixmap(pm)
-            try:
-                self._extra_fetch_cb(url, _on_px)
-            except Exception:
-                self._drop_bad_url(url)
-
-    def _drop_bad_url(self, url: str):
-        """Remove a URL that failed to load from the cycling list."""
-        try:
-            i = self._screenshots.index(url)
-        except ValueError:
-            return
-        del self._screenshots[i]
-        self._shot_cache.pop(url, None)
-        if not self._screenshots:
-            self._timer.stop()
-            self._shot_index = 0
-            self._update_nav()
-            self._render_placeholder()
-            return
-        # Pick a new index that still points to a valid entry.
-        new_idx = self._shot_index
-        if i < self._shot_index:
-            new_idx = self._shot_index - 1
-        if new_idx >= len(self._screenshots):
-            new_idx = 0
-        self._shot_index = new_idx
-        self._update_nav()
-        if len(self._screenshots) <= 1:
-            self._timer.stop()
-        self._show_index(new_idx)
-
-    def _display_pixmap(self, pm: QPixmap):
-        if pm is None or pm.isNull():
-            return
-        sz = self._img_lbl.size()
-        self._img_lbl.setPixmap(
-            pm.scaled(sz, Qt.KeepAspectRatio, Qt.SmoothTransformation)
-        )
-        self._img_lbl.setText("")
-        self._position_tag_overlay()
-
-    def resizeEvent(self, ev):
-        super().resizeEvent(ev)
-        if self._screenshots and self._shot_index < len(self._screenshots):
-            cached = self._shot_cache.get(self._screenshots[self._shot_index])
-            if cached:
-                self._display_pixmap(cached)
-        elif not self._screenshots and self._placeholder_label:
-            self._render_placeholder()
-
-    def hideEvent(self, ev):
-        self._timer.stop()
-        super().hideEvent(ev)
-
-
-def _gallery_viewer_refresh_meta(viewer: "GalleryItemViewer",
-                                  title: str, rows: list):
-    """Rebuild the metadata scroll area of an already-shown GalleryItemViewer.
-
-    Delegates to viewer.refresh_meta() which handles the widget swap cleanly.
-    Called from async callbacks after the viewer is already embedded in the pane.
-    """
-    try:
-        viewer.refresh_meta(title, rows)
-    except Exception:
-        pass
-
-
-class BackgroundWidget(QWidget):
-    """A QWidget that paints a chosen (or randomly cycling) image from the
-    same directory as the script, scaled to fill the entire widget area, blended at
-    a configurable opacity level (0–100 %, default 45 %).
-
-    Modes
-    -----
-    Random (default): cycles through all available background images every
-    BG_CYCLE_INTERVAL_MS milliseconds using a QTimer.
-    Fixed: a specific image path is set via set_bg_image(path); the timer stops.
-    """
-
-    DEFAULT_OPACITY      = 45
-    BG_CYCLE_INTERVAL_MS = 5000   # 5 seconds
-
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-        self._bg_pixmap   = None
-        self._bg_opacity  = self.DEFAULT_OPACITY  # percent 0-100
-        self._bg_paths    = []                     # all discovered image paths
-        self._bg_index    = -1                     # current index for cycling
-        self._bg_fixed    = False                  # True = specific image locked
-        self._cycle_timer = QTimer(self)
-        self._cycle_timer.setInterval(self.BG_CYCLE_INTERVAL_MS)
-        self._cycle_timer.timeout.connect(self._cycle_next)
-        # Start in random-cycling mode immediately
-        self._bg_paths = self._discover_backgrounds()
-        if self._bg_paths:
-            import random
-            self._bg_index = random.randrange(len(self._bg_paths))
-            self._bg_pixmap = self._load_path(self._bg_paths[self._bg_index])
-            self._cycle_timer.start()
-
-    # ------------------------------------------------------------------
-    # Public API
-    # ------------------------------------------------------------------
-
-    def set_bg_opacity(self, percent: int):
-        """Set background image opacity (0 = invisible, 100 = fully opaque)."""
-        self._bg_opacity = max(0, min(100, int(percent)))
-        self.update()
-
-    def set_bg_image(self, path: str):
-        """Switch to a fixed image (non-empty path) or back to random cycling
-        (empty / None path).  *path* should be an absolute file path."""
-        if path:
-            self._bg_fixed = True
-            self._cycle_timer.stop()
-            self._bg_pixmap = self._load_path(path)
-            self.update()
-        else:
-            self._bg_fixed = False
-            self._bg_paths = self._discover_backgrounds()
-            if self._bg_paths:
-                import random
-                self._bg_index = random.randrange(len(self._bg_paths))
-                self._bg_pixmap = self._load_path(self._bg_paths[self._bg_index])
-                self._cycle_timer.start()
-            else:
-                self._bg_pixmap = None
-                self.update()
-
-    def bg_paths(self) -> list:
-        """Return the list of discovered background image paths (may be empty)."""
-        return list(self._bg_paths)
-
-    # ------------------------------------------------------------------
-    # Internal helpers
-    # ------------------------------------------------------------------
-
-    @staticmethod
-    def _discover_backgrounds() -> list:
-        """Return a list of image paths from the script directory first,
-        then from Qt embedded resources (rc_backgrounds)."""
-        image_extensions = {".png", ".jpg", ".jpeg", ".bmp", ".gif", ".webp"}
-        paths = []
-
-        # Filesystem images (same directory as the running script/exe)
-        bg_dir = os.path.dirname(os.path.abspath(sys.argv[0]))
-        if os.path.isdir(bg_dir):
-            paths.extend(sorted(
-                os.path.join(bg_dir, f)
-                for f in os.listdir(bg_dir)
-                if os.path.splitext(f)[1].lower() in image_extensions
-            ))
-
-        # Qt resource images (embedded via rc_backgrounds)
-        from PySide6.QtCore import QDir as _QDir
-        for _name in _QDir(":/").entryList():
-            if os.path.splitext(_name)[1].lower() in image_extensions:
-                rc_path = ":/" + _name
-                if rc_path not in paths:
-                    paths.append(rc_path)
-
-        return paths
-
-    @staticmethod
-    def _load_path(path: str):
-        if not path:
-            return None
-        px = QPixmap(path)
-        return px if not px.isNull() else None
-
-    def _cycle_next(self):
-        """Advance to the next image in the rotation."""
-        if not self._bg_paths:
-            return
-        self._bg_index = (self._bg_index + 1) % len(self._bg_paths)
-        self._bg_pixmap = self._load_path(self._bg_paths[self._bg_index])
-        self.update()
-
-    # kept for backward-compat (not used internally anymore)
-    @staticmethod
-    def _load_random_background():
-        import random
-        bg_dir = os.path.dirname(os.path.abspath(sys.argv[0]))
-        if not os.path.isdir(bg_dir):
-            return None
-        image_extensions = {".png", ".jpg", ".jpeg", ".bmp", ".gif", ".webp"}
-        candidates = [
-            os.path.join(bg_dir, f)
-            for f in os.listdir(bg_dir)
-            if os.path.splitext(f)[1].lower() in image_extensions
-        ]
-        if not candidates:
-            return None
-        chosen = random.choice(candidates)
-        px = QPixmap(chosen)
-        return px if not px.isNull() else None
-
-    def paintEvent(self, event):
-        super().paintEvent(event)
-        if self._bg_pixmap and self._bg_opacity > 0:
-            painter = QPainter(self)
-            painter.setOpacity(self._bg_opacity / 100.0)
-            painter.drawPixmap(self.rect(), self._bg_pixmap)
 
 def _apply_completer_fix_to_children(widget: QWidget):
     for child in widget.findChildren(QWidget):
