+++ zxnu_workers.py
@@ -1,0 +1,196 @@
+"""Background worker, signal and progress-dialog classes for zx-next-unite.
+
+Extracted from zx-next-unite.py."""
+
+import threading
+from PySide6.QtCore import QObject, QRunnable, QSortFilterProxyModel, QTimer, Qt, Signal, Slot
+from PySide6.QtGui import QFontInfo
+from PySide6.QtWidgets import QDialog, QHBoxLayout, QLabel, QProgressBar, QPushButton, QVBoxLayout
+
+
+class DotDotFirstProxyModel(QSortFilterProxyModel):
+    """Proxy model that always keeps the '..' parent directory entry at the top."""
+    def lessThan(self, left, right):
+        left_name = self.sourceModel().fileName(left)
+        right_name = self.sourceModel().fileName(right)
+        if left_name == "..":
+            return True
+        if right_name == "..":
+            return False
+        return super().lessThan(left, right)
+
+    def filterAcceptsRow(self, source_row, source_parent):
+        source_model = self.sourceModel()
+        index = source_model.index(source_row, 0, source_parent)
+        # Always show the parent-directory entry
+        if source_model.fileName(index) == "..":
+            return True
+        pattern = self.filterRegularExpression().pattern()
+        if not pattern:
+            return True
+        name = source_model.fileName(index)
+        return pattern.lower() in name.lower()
+
+class WorkerSignals(QObject):
+
+    finished = Signal()
+    error = Signal(tuple)
+    result = Signal(object)
+    progress = Signal(int)
+
+
+class NextSyncSignals(QObject):
+    """Signals used to marshal nextsync progress back to the main thread."""
+    progress = Signal(int)   # 0-100 per-file progress
+    status   = Signal(str)   # single-line status message
+    finished = Signal()      # emitted when the job thread exits
+    cancelled = Signal()     # emitted when job stopped due to cancel request
+
+
+class HdfTaskSignals(QObject):
+    """Signals for background hdfmonkey task workers."""
+    progress  = Signal(int)   # 0-100
+    status    = Signal(str)   # "action line\nfilename line"
+    finished  = Signal()
+    error     = Signal(str)   # human-readable error message
+    cancelled = Signal()      # emitted when the worker stopped early due to cancel
+
+
+class HdfTaskWorker(QRunnable):
+    """Generic QRunnable that runs a callable on the thread pool.
+    The callable receives (signals, cancel_event, *args, **kwargs).
+    Call worker.cancel() from the UI thread to request early termination."""
+
+    def __init__(self, fn, *args, **kwargs):
+        super().__init__()
+        self.fn           = fn
+        self.args         = args
+        self.kwargs       = kwargs
+        self.signals      = HdfTaskSignals()
+        self.cancel_event = threading.Event()
+        self.setAutoDelete(True)
+
+    def cancel(self):
+        self.cancel_event.set()
+
+    @Slot()
+    def run(self):
+        try:
+            self.fn(self.signals, self.cancel_event, *self.args, **self.kwargs)
+        except Exception as exc:
+            self.signals.error.emit(str(exc))
+        finally:
+            if self.cancel_event.is_set():
+                self.signals.cancelled.emit()
+            self.signals.finished.emit()
+
+
+class HdfProgressDialog(QDialog):
+    """Modal progress dialog with live status, progress bar, spinner, and Cancel button."""
+
+    cancel_requested = Signal()
+
+    def __init__(self, title, parent=None):
+        super().__init__(parent)
+        self.setWindowTitle(title)
+        self.setModal(True)
+        self.setMinimumWidth(540)
+        self.setWindowFlags(self.windowFlags() & ~Qt.WindowCloseButtonHint)
+
+        layout = QVBoxLayout(self)
+        layout.setSpacing(8)
+
+        # Spinner + action label on one row
+        action_row = QHBoxLayout()
+        self._spinner_label = QLabel("")
+        self._spinner_label.setFixedWidth(22)
+        action_row.addWidget(self._spinner_label)
+        self._action_label = QLabel("Starting\u2026")
+        self._action_label.setWordWrap(True)
+        action_row.addWidget(self._action_label, 1)
+        layout.addLayout(action_row)
+
+        # Current filename (smaller, muted)
+        self._file_label = QLabel("")
+        self._file_label.setWordWrap(True)
+        _font = self._file_label.font()
+        _ps = _font.pointSize()
+        if _ps <= 0:
+            _ps = max(QFontInfo(_font).pointSize(), 9)
+        _font.setPointSize(max(_ps - 1, 8))
+        self._file_label.setFont(_font)
+        layout.addWidget(self._file_label)
+
+        # Progress bar
+        self._bar = QProgressBar()
+        self._bar.setRange(0, 100)
+        self._bar.setValue(0)
+        self._bar.setTextVisible(True)
+        layout.addWidget(self._bar)
+
+        # Cancel button
+        btn_row = QHBoxLayout()
+        btn_row.addStretch()
+        self._cancel_btn = QPushButton("Cancel")
+        self._cancel_btn.setFixedWidth(90)
+        self._cancel_btn.clicked.connect(self._on_cancel_clicked)
+        btn_row.addWidget(self._cancel_btn)
+        layout.addLayout(btn_row)
+
+        self._cancelled = False
+        self._spinner_frames = ["\u25f4", "\u25f7", "\u25f6", "\u25f5"]
+        self._spinner_idx    = 0
+
+        self._anim_timer = QTimer(self)
+        self._anim_timer.setInterval(120)
+        self._anim_timer.timeout.connect(self._tick_spinner)
+        self._anim_timer.start()
+
+    # ------------------------------------------------------------------
+    @Slot()
+    def _on_cancel_clicked(self):
+        self._cancelled = True
+        self._cancel_btn.setEnabled(False)
+        self._action_label.setText("Cancelling\u2026")
+        self._file_label.setText("")
+        self.cancel_requested.emit()
+
+    @Slot()
+    def _tick_spinner(self):
+        self._spinner_idx = (self._spinner_idx + 1) % len(self._spinner_frames)
+        self._spinner_label.setText(self._spinner_frames[self._spinner_idx])
+
+    @Slot(int)
+    def set_progress(self, value: int):
+        """value == -1 activates the indeterminate (busy) marquee animation."""
+        if value < 0:
+            self._bar.setRange(0, 0)   # Qt marquee mode
+        else:
+            if self._bar.maximum() == 0:
+                self._bar.setRange(0, 100)
+            self._bar.setValue(value)
+
+    @Slot(str)
+    def set_status(self, text: str):
+        """Expects 'Action description\nFilename or detail'."""
+        if self._cancelled:
+            return
+        lines = text.split("\n", 1)
+        self._action_label.setText(lines[0])
+        self._file_label.setText(lines[1] if len(lines) > 1 else "")
+
+    @Slot()
+    def mark_cancelled(self):
+        """Called when the worker confirms it stopped early."""
+        self._action_label.setText("Cancelled.")
+        self._file_label.setText("")
+
+    def closeEvent(self, event):
+        self._anim_timer.stop()
+        super().closeEvent(event)
+
+
+# Export every public/private module-level name (including the
+# underscore-prefixed helpers and caches) so `from <module> import *`
+# in the main file picks them all up.
+__all__ = [_n for _n in dir() if not _n.startswith('__')]
